+++ resources/views/admin/chat/index.blade.php
@@ -1,1579 +1,729 @@
 @extends('layouts.admin.contentLayoutMaster')
 
-@section('title', 'Quản lý Chat - Admin')
+@section('title', 'Super Admin Dashboard')
 
 @section('content')
     <style>
-        /* Dark mode variables */
         :root {
-            --background: 0 0% 100%;
-            --foreground: 222.2 84% 4.9%;
-            --card: 0 0% 100%;
-            --card-foreground: 222.2 84% 4.9%;
-            --popover: 0 0% 100%;
-            --popover-foreground: 222.2 84% 4.9%;
-            --primary: 221.2 83.2% 53.3%;
-            --primary-foreground: 210 40% 98%;
-            --secondary: 210 40% 96%;
-            --secondary-foreground: 222.2 84% 4.9%;
-            --muted: 210 40% 96%;
-            --muted-foreground: 215.4 16.3% 46.9%;
-            --accent: 210 40% 96%;
-            --accent-foreground: 222.2 84% 4.9%;
-            --destructive: 0 84.2% 60.2%;
-            --destructive-foreground: 210 40% 98%;
-            --border: 214.3 31.8% 91.4%;
-            --input: 214.3 31.8% 91.4%;
-            --ring: 221.2 83.2% 53.3%;
-            --radius: 0.5rem;
-        }
-
-        .dark {
-            --background: 222.2 84% 4.9%;
-            --foreground: 210 40% 98%;
-            --card: 222.2 84% 4.9%;
-            --card-foreground: 210 40% 98%;
-            --popover: 222.2 84% 4.9%;
-            --popover-foreground: 210 40% 98%;
-            --primary: 217.2 91.2% 59.8%;
-            --primary-foreground: 222.2 84% 4.9%;
-            --secondary: 217.2 32.6% 17.5%;
-            --secondary-foreground: 210 40% 98%;
-            --muted: 217.2 32.6% 17.5%;
-            --muted-foreground: 215 20.2% 65.1%;
-            --accent: 217.2 32.6% 17.5%;
-            --accent-foreground: 210 40% 98%;
-            --destructive: 0 62.8% 30.6%;
-            --destructive-foreground: 210 40% 98%;
-            --border: 217.2 32.6% 17.5%;
-            --input: 217.2 32.6% 17.5%;
-            --ring: 224.3 76.3% 94.1%;
-        }
-
-        * {
-            border-color: hsl(var(--border));
+            --primary-color: #3b82f6;
+            --primary-hover: #2563eb;
+            --text-primary: #111827;
+            --text-secondary: #4b5563;
+            --text-muted: #6b7280;
+            --bg-hover: #f3f4f6;
+            --border-color: #e5e7eb;
+            --badge-waiting: #fef3c7;
+            --badge-waiting-text: #92400e;
+            --badge-high: #fee2e2;
+            --badge-high-text: #991b1b;
+            --badge-distributed: #dbeafe;
+            --badge-distributed-text: #1e40af;
         }
 
         body {
-            background-color: hsl(var(--background));
-            color: hsl(var(--foreground));
+            font-family: 'Inter', -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+            color: var(--text-primary);
+            background-color: #f9fafb;
         }
 
-        .bg-background {
-            background-color: hsl(var(--background));
+        .chat-container {
+            display: flex;
+            height: calc(100vh - 80px);
+            background-color: white;
+            border-radius: 8px;
+            overflow: hidden;
+            box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+            border: 1px solid var(--border-color);
         }
 
-        .bg-foreground {
-            background-color: hsl(var(--foreground));
+        .chat-sidebar {
+            width: 380px;
+            border-right: 1px solid var(--border-color);
+            display: flex;
+            flex-direction: column;
+            background-color: #f9fafb;
         }
 
-        .bg-card {
-            background-color: hsl(var(--card));
+        .chat-sidebar-header {
+            padding: 20px;
+            border-bottom: 1px solid var(--border-color);
+            background-color: white;
         }
 
-        .bg-card-foreground {
-            background-color: hsl(var(--card-foreground));
+        .chat-sidebar-header h2 {
+            margin: 0;
+            font-size: 1.25rem;
+            font-weight: 600;
         }
 
-        .bg-popover {
-            background-color: hsl(var(--popover));
+        .chat-sidebar-header p {
+            margin: 4px 0 0;
+            color: var(--text-secondary);
+            font-size: 0.875rem;
         }
 
-        .bg-popover-foreground {
-            background-color: hsl(var(--popover-foreground));
+        .chat-list {
+            flex: 1;
+            overflow-y: auto;
         }
 
-        .bg-primary {
-            background-color: hsl(var(--primary));
+        .chat-item {
+            padding: 16px 20px;
+            border-bottom: 1px solid var(--border-color);
+            cursor: pointer;
+            transition: background-color 0.2s;
+            position: relative;
+            background-color: white;
         }
 
-        .bg-primary-foreground {
-            background-color: hsl(var(--primary-foreground));
-        }
-
-        .bg-secondary {
-            background-color: hsl(var(--secondary));
-        }
-
-        .bg-secondary-foreground {
-            background-color: hsl(var(--secondary-foreground));
-        }
-
-        .bg-muted {
-            background-color: hsl(var(--muted));
-        }
-
-        .bg-muted-foreground {
-            background-color: hsl(var(--muted-foreground));
-        }
-
-        .bg-accent {
-            background-color: hsl(var(--accent));
-        }
-
-        .bg-accent-foreground {
-            background-color: hsl(var(--accent-foreground));
-        }
-
-        .bg-destructive {
-            background-color: hsl(var(--destructive));
-        }
-
-        .bg-destructive-foreground {
-            background-color: hsl(var(--destructive-foreground));
-        }
-
-        .text-background {
-            color: hsl(var(--background));
-        }
-
-        .text-foreground {
-            color: hsl(var(--foreground));
-        }
-
-        .text-card {
-            color: hsl(var(--card));
-        }
-
-        .text-card-foreground {
-            color: hsl(var(--card-foreground));
-        }
-
-        .text-popover {
-            color: hsl(var(--popover));
-        }
-
-        .text-popover-foreground {
-            color: hsl(var(--popover-foreground));
-        }
-
-        .text-primary {
-            color: hsl(var(--primary));
-        }
-
-        .text-primary-foreground {
-            color: hsl(var(--primary-foreground));
-        }
-
-        .text-secondary {
-            color: hsl(var(--secondary));
-        }
-
-        .text-secondary-foreground {
-            color: hsl(var(--secondary-foreground));
-        }
-
-        .text-muted {
-            color: hsl(var(--muted));
-        }
-
-        .text-muted-foreground {
-            color: hsl(var(--muted-foreground));
-        }
-
-        .text-accent {
-            color: hsl(var(--accent));
-        }
-
-        .text-accent-foreground {
-            color: hsl(var(--accent-foreground));
-        }
-
-        .text-destructive {
-            color: hsl(var(--destructive));
-        }
-
-        .text-destructive-foreground {
-            color: hsl(var(--destructive-foreground));
-        }
-
-        .border {
-            border-color: hsl(var(--border));
-        }
-
-        .border-border {
-            border-color: hsl(var(--border));
-        }
-
-        .border-input {
-            border-color: hsl(var(--input));
-        }
-
-        /* Custom scrollbar */
-        .custom-scrollbar::-webkit-scrollbar {
-            width: 6px;
-        }
-
-        .custom-scrollbar::-webkit-scrollbar-track {
-            background: hsl(var(--muted));
-        }
-
-        .custom-scrollbar::-webkit-scrollbar-thumb {
-            background: hsl(var(--muted-foreground));
-            border-radius: 3px;
-        }
-
-        .custom-scrollbar::-webkit-scrollbar-thumb:hover {
-            background: hsl(var(--accent-foreground));
-        }
-
-        /* Chat specific styles */
         .chat-item:hover {
-            background-color: hsl(var(--accent));
+            background-color: var(--bg-hover);
         }
 
         .chat-item.active {
-            background-color: hsl(var(--primary) / 0.1);
-            border-left: 4px solid hsl(var(--primary));
+            background-color: #f0f9ff;
+            border-left: 3px solid var(--primary-color);
         }
 
-        /* Message bubbles */
-        .message-admin {
-            background-color: hsl(var(--primary));
-            color: hsl(var(--primary-foreground));
+        .chat-item-header {
+            display: flex;
+            justify-content: space-between;
+            align-items: flex-start;
+            margin-bottom: 4px;
         }
 
-        .message-customer {
-            background-color: hsl(var(--card));
-            color: hsl(var(--card-foreground));
-            border: 1px solid hsl(var(--border));
+        .chat-item-name {
+            font-weight: 600;
+            font-size: 0.9375rem;
+            margin: 0;
         }
 
-        /* Quick reply buttons */
-        .quick-reply-btn {
-            background-color: hsl(var(--muted));
-            color: hsl(var(--muted-foreground));
-            transition: all 0.2s ease;
+        .chat-item-time {
+            font-size: 0.75rem;
+            color: var(--text-muted);
         }
 
-        .quick-reply-btn:hover {
-            background-color: hsl(var(--accent));
-            color: hsl(var(--accent-foreground));
+        .chat-item-email {
+            font-size: 0.8125rem;
+            color: var(--text-secondary);
+            margin: 0 0 8px;
         }
 
-        /* Status indicators */
-        .status-online {
-            background-color: #22c55e;
+        .chat-item-preview {
+            font-size: 0.875rem;
+            color: var(--text-secondary);
+            margin: 0 0 8px;
+            white-space: nowrap;
+            overflow: hidden;
+            text-overflow: ellipsis;
         }
 
-        .status-offline {
-            background-color: #6b7280;
+        .chat-item-footer {
+            display: flex;
+            justify-content: space-between;
+            align-items: center;
         }
 
-        /* Theme toggle button */
-        .theme-toggle {
-            position: relative;
-            width: 44px;
+        .chat-item-badges {
+            display: flex;
+            gap: 8px;
+        }
+
+        .badge {
+            font-size: 0.75rem;
+            padding: 4px 10px;
+            border-radius: 9999px;
+            font-weight: 500;
+        }
+
+        .badge-waiting {
+            background-color: var(--badge-waiting);
+            color: var(--badge-waiting-text);
+        }
+
+        .badge-high {
+            background-color: var(--badge-high);
+            color: var(--badge-high-text);
+        }
+
+        .badge-distributed {
+            background-color: var(--badge-distributed);
+            color: var(--badge-distributed-text);
+        }
+
+        .unread-badge {
+            background-color: #ef4444;
+            color: white;
+            width: 24px;
             height: 24px;
-            background-color: hsl(var(--muted));
-            border-radius: 12px;
-            transition: background-color 0.3s ease;
-            cursor: pointer;
-            border: 1px solid hsl(var(--border));
-        }
-
-        .theme-toggle.dark {
-            background-color: hsl(var(--primary));
-        }
-
-        .theme-toggle-handle {
-            position: absolute;
-            top: 2px;
-            left: 2px;
-            width: 18px;
-            height: 18px;
-            background-color: hsl(var(--background));
             border-radius: 50%;
-            transition: transform 0.3s ease;
             display: flex;
             align-items: center;
             justify-content: center;
-            font-size: 10px;
+            font-size: 0.75rem;
+            font-weight: 600;
+            position: absolute;
+            top: 16px;
+            right: 20px;
         }
 
-        .theme-toggle.dark .theme-toggle-handle {
-            transform: translateX(20px);
+        .chat-main {
+            flex: 1;
+            display: flex;
+            flex-direction: column;
         }
 
-        /* Input focus styles */
-        input:focus,
-        textarea:focus,
-        select:focus {
-            outline: none;
-            border-color: hsl(var(--primary));
-            box-shadow: 0 0 0 2px hsl(var(--primary) / 0.2);
+        .chat-header {
+            padding: 16px 24px;
+            border-bottom: 1px solid var(--border-color);
+            display: flex;
+            justify-content: space-between;
+            align-items: center;
         }
 
-        /* Button styles */
-        .btn {
-            display: inline-flex;
+        .chat-header-user {
+            display: flex;
+            align-items: center;
+            gap: 12px;
+        }
+
+        .chat-avatar {
+            width: 40px;
+            height: 40px;
+            border-radius: 50%;
+            background-color: #e5e7eb;
+            display: flex;
             align-items: center;
             justify-content: center;
-            padding: 0.5rem 1rem;
-            border-radius: 0.375rem;
+            font-weight: 600;
+            font-size: 1rem;
+        }
+
+        .chat-header-info h3 {
+            margin: 0;
+            font-size: 1rem;
+            font-weight: 600;
+        }
+
+        .chat-header-info p {
+            margin: 2px 0 0;
             font-size: 0.875rem;
+            color: var(--text-secondary);
+        }
+
+        .chat-header-actions {
+            display: flex;
+            gap: 8px;
+        }
+
+        .chat-messages {
+            flex: 1;
+            padding: 24px;
+            overflow-y: auto;
+            background-color: #f9fafb;
+        }
+
+        .message-group {
+            margin-bottom: 24px;
+        }
+
+        .message-sender {
+            display: flex;
+            align-items: center;
+            gap: 8px;
+            margin-bottom: 4px;
+        }
+
+        .message-sender-name {
             font-weight: 500;
-            transition: all 0.2s ease;
+            font-size: 0.875rem;
+        }
+
+        .message-sender-type {
+            font-size: 0.75rem;
+            color: var(--text-muted);
+        }
+
+        .message-time {
+            font-size: 0.75rem;
+            color: var(--text-muted);
+            margin-left: 8px;
+        }
+
+        .message-bubble {
+            padding: 12px 16px;
+            border-radius: 8px;
+            max-width: 80%;
+            font-size: 0.9375rem;
+            line-height: 1.5;
+        }
+
+        .message-customer {
+            background-color: #f3f4f6;
+            color: var(--text-primary);
+            align-self: flex-start;
+            border-bottom-left-radius: 2px;
+        }
+
+        .message-admin {
+            background-color: var(--primary-color);
+            color: white;
+            align-self: flex-end;
+            border-bottom-right-radius: 2px;
+            margin-left: auto;
+        }
+
+        .chat-input-container {
+            padding: 16px 24px;
+            border-top: 1px solid var(--border-color);
+            display: flex;
+            align-items: center;
+            gap: 12px;
+        }
+
+        .chat-input {
+            flex: 1;
+            border: 1px solid var(--border-color);
+            border-radius: 8px;
+            padding: 12px 16px;
+            font-size: 0.9375rem;
+            resize: none;
+            height: 48px;
+            max-height: 120px;
+            overflow-y: auto;
+        }
+
+        .chat-input:focus {
+            outline: none;
+            border-color: var(--primary-color);
+            box-shadow: 0 0 0 2px rgba(59, 130, 246, 0.1);
+        }
+
+        .chat-send-btn {
+            width: 48px;
+            height: 48px;
+            border-radius: 8px;
+            background-color: var(--primary-color);
+            color: white;
+            border: none;
             cursor: pointer;
-            border: 1px solid transparent;
+            display: flex;
+            align-items: center;
+            justify-content: center;
+            transition: background-color 0.2s;
         }
 
-        .btn-primary {
-            background-color: hsl(var(--primary));
-            color: hsl(var(--primary-foreground));
+        .chat-send-btn:hover {
+            background-color: var(--primary-hover);
         }
 
-        .btn-primary:hover {
-            background-color: hsl(var(--primary) / 0.9);
+        .chat-send-btn:disabled {
+            background-color: #e5e7eb;
+            cursor: not-allowed;
         }
 
-        .btn-secondary {
-            background-color: hsl(var(--secondary));
-            color: hsl(var(--secondary-foreground));
+        .chat-tools-btn {
+            width: 48px;
+            height: 48px;
+            border-radius: 8px;
+            background-color: white;
+            color: var(--text-secondary);
+            border: 1px solid var(--border-color);
+            cursor: pointer;
+            display: flex;
+            align-items: center;
+            justify-content: center;
+            transition: all 0.2s;
         }
 
-        .btn-secondary:hover {
-            background-color: hsl(var(--secondary) / 0.8);
+        .chat-tools-btn:hover {
+            background-color: var(--bg-hover);
+            color: var(--text-primary);
         }
 
-        .btn-outline {
-            border: 1px solid hsl(var(--border));
-            background-color: hsl(var(--background));
-            color: hsl(var(--foreground));
+        .distribution-select {
+            width: 100%;
+            padding: 10px 12px;
+            border-radius: 6px;
+            border: 1px solid var(--border-color);
+            font-size: 0.875rem;
+            color: var(--text-primary);
+            background-color: white;
+            margin-top: 16px;
+            appearance: none;
+            background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' fill='none' viewBox='0 0 24 24' stroke='%236b7280'%3E%3Cpath stroke-linecap='round' stroke-linejoin='round' stroke-width='2' d='M19 9l-7 7-7-7'%3E%3C/path%3E%3C/svg%3E");
+            background-repeat: no-repeat;
+            background-position: right 12px center;
+            background-size: 16px;
+            padding-right: 36px;
         }
 
-        .btn-outline:hover {
-            background-color: hsl(var(--accent));
-            color: hsl(var(--accent-foreground));
+        .distribution-select:focus {
+            outline: none;
+            border-color: var(--primary-color);
+            box-shadow: 0 0 0 2px rgba(59, 130, 246, 0.1);
         }
 
-        .btn-destructive {
-            background-color: hsl(var(--destructive));
-            color: hsl(var(--destructive-foreground));
+        .empty-state {
+            display: flex;
+            flex-direction: column;
+            align-items: center;
+            justify-content: center;
+            height: 100%;
+            color: var(--text-muted);
+            text-align: center;
+            padding: 2rem;
         }
 
-        .btn-destructive:hover {
-            background-color: hsl(var(--destructive) / 0.9);
+        .empty-state i {
+            font-size: 3rem;
+            margin-bottom: 1rem;
+            color: #d1d5db;
         }
 
-        /* Card styles */
-        .card {
-            background-color: hsl(var(--card));
-            color: hsl(var(--card-foreground));
-            border: 1px solid hsl(var(--border));
-            border-radius: 0.5rem;
+        .empty-state h3 {
+            font-size: 1.25rem;
+            font-weight: 500;
+            margin-bottom: 0.5rem;
+            color: var(--text-primary);
         }
 
-        /* Modal styles */
-        .modal-overlay {
-            background-color: rgba(0, 0, 0, 0.5);
+        .empty-state p {
+            font-size: 0.875rem;
         }
 
-        .modal-content {
-            background-color: hsl(var(--card));
-            color: hsl(var(--card-foreground));
-            border: 1px solid hsl(var(--border));
+        .loading-spinner {
+            display: flex;
+            align-items: center;
+            justify-content: center;
+            padding: 2rem;
+        }
+
+        .loading-spinner i {
+            font-size: 2rem;
+            color: var(--primary-color);
+        }
+
+        /* Custom scrollbar */
+        .chat-list::-webkit-scrollbar,
+        .chat-messages::-webkit-scrollbar,
+        .chat-input::-webkit-scrollbar {
+            width: 6px;
+        }
+
+        .chat-list::-webkit-scrollbar-track,
+        .chat-messages::-webkit-scrollbar-track,
+        .chat-input::-webkit-scrollbar-track {
+            background: transparent;
+        }
+
+        .chat-list::-webkit-scrollbar-thumb,
+        .chat-messages::-webkit-scrollbar-thumb,
+        .chat-input::-webkit-scrollbar-thumb {
+            background-color: #d1d5db;
+            border-radius: 20px;
+        }
+
+        .chat-list::-webkit-scrollbar-thumb:hover,
+        .chat-messages::-webkit-scrollbar-thumb:hover,
+        .chat-input::-webkit-scrollbar-thumb:hover {
+            background-color: #9ca3af;
+        }
+
+        .stats-container {
+            grid-template-columns: 1fr;
+            gap: 1rem;
+        }
+
+        @media (max-width: 1024px) {
+            .stats-container {
+                grid-template-columns: repeat(2, 1fr);
+            }
+        }
+
+        .stats-container {
+            display: grid;
+            grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
+            gap: 1.5rem;
+            margin-bottom: 2rem;
+        }
+
+        .stat-card .stat-header {
+            display: flex;
+            justify-content: space-between;
+            align-items: flex-start;
+            margin-bottom: 1rem;
+        }
+
+        .stat-card .stat-icon {
+            width: 48px;
+            height: 48px;
+            border-radius: 12px;
+            display: flex;
+            align-items: center;
+            justify-content: center;
+            font-size: 1.25rem;
+            color: white;
+        }
+
+        .stat-card.new .stat-icon {
+            background: linear-gradient(135deg, var(--primary-color), var(--primary-hover));
+        }
+
+        .stat-card.active .stat-icon {
+            background: linear-gradient(135deg, var(--success-color), #059669);
+        }
+
+        .stat-card.priority .stat-icon {
+            background: linear-gradient(135deg, var(--warning-color), #d97706);
+        }
+
+        .stat-card.total .stat-icon {
+            background: linear-gradient(135deg, var(--danger-color), #dc2626);
+        }
+
+        .stat-card .stat-number {
+            font-size: 2rem;
+            font-weight: 800;
+            color: var(--text-primary);
+            margin-bottom: 0.5rem;
+            line-height: 1;
+        }
+
+        .stat-card .stat-label {
+            color: var(--text-secondary);
+            font-size: 0.875rem;
+            font-weight: 600;
+            text-transform: uppercase;
+            letter-spacing: 0.5px;
+        }
+
+        .stat-card .stat-trend {
+            font-size: 0.75rem;
+            color: var(--success-color);
+            font-weight: 600;
+            margin-top: 0.5rem;
         }
     </style>
 
-    <div class="min-h-screen bg-background">
-        <!-- Header -->
-        <div class="bg-card shadow-sm border-b border-border">
-            <div class="px-6 py-4">
-                <div class="flex items-center justify-between">
-                    <div>
-                        <h1 class="text-2xl font-bold text-foreground">Quản lý Chat</h1>
-                        <p class="text-muted-foreground">Hỗ trợ khách hàng trực tuyến</p>
-                    </div>
-                    <div class="flex items-center gap-4">
-                        <!-- Theme Toggle -->
-                        <div class="flex items-center gap-2">
-                            <span class="text-sm text-muted-foreground">Theme:</span>
-                            <button id="themeToggle" class="theme-toggle">
-                                <div class="theme-toggle-handle">
-                                    <span id="themeIcon">🌙</span>
-                                </div>
-                            </button>
-                        </div>
+    <meta name="csrf-token" content="{{ csrf_token() }}">
 
-                        <!-- Admin Status Toggle -->
-                        <div class="flex items-center gap-2">
-                            <span class="text-sm text-muted-foreground">Trạng thái:</span>
-                            <button id="statusToggle"
-                                class="relative inline-flex h-6 w-11 items-center rounded-full bg-green-500 transition-colors">
-                                <span
-                                    class="inline-block h-4 w-4 transform rounded-full bg-white transition-transform translate-x-6"></span>
-                            </button>
-                            <span id="statusText" class="text-sm font-medium text-green-600">Online</span>
-                        </div>
+    <div id="chat-container" data-conversation-id="{{ $selectedConversation->id ?? '' }}" data-user-id="{{ auth()->id() }}"
+        data-user-type="admin">
+        <div id="chat-messages" class="chat-messages"></div>
+        <div class="chat-container">
+            <!-- Chat Sidebar -->
 
-                        <!-- Notifications -->
-                        <button class="relative p-2 text-muted-foreground hover:text-foreground hover:bg-accent rounded-lg">
-                            <i class="fas fa-bell text-lg"></i>
-                            <span id="notificationBadge"
-                                class="absolute -top-1 -right-1 h-5 w-5 bg-red-500 text-white text-xs rounded-full flex items-center justify-center hidden">0</span>
-                        </button>
-                    </div>
-                </div>
-            </div>
-        </div>
+            <div class="chat-sidebar">
 
-        <div class="flex h-[calc(100vh-120px)]">
-            <!-- Chat List Sidebar -->
-            <div class="w-80 bg-card border-r border-border flex flex-col">
-                <!-- Search & Filter -->
-                <div class="p-4 border-b border-border">
-                    <div class="relative mb-3">
-                        <input type="text" id="searchChats" placeholder="Tìm kiếm cuộc trò chuyện..."
-                            class="w-full pl-10 pr-4 py-2 border border-input rounded-lg bg-background text-foreground focus:ring-2 focus:ring-primary focus:border-primary">
-                        <i
-                            class="fas fa-search absolute left-3 top-1/2 transform -translate-y-1/2 text-muted-foreground"></i>
-                    </div>
-
-                    <div class="flex gap-2">
-                        <select id="filterStatus"
-                            class="flex-1 text-sm border border-input rounded-lg px-3 py-1 bg-background text-foreground focus:ring-2 focus:ring-primary">
-                            <option value="all">Tất cả</option>
-                            <option value="active">Đang hoạt động</option>
-                            <option value="waiting">Chờ phản hồi</option>
-                            <option value="closed">Đã đóng</option>
-                        </select>
-                        <button id="refreshChats"
-                            class="px-3 py-1 text-muted-foreground hover:text-foreground hover:bg-accent rounded-lg">
-                            <i class="fas fa-sync-alt"></i>
-                        </button>
-                    </div>
+                <div class="chat-sidebar-header">
+                    <h2>Super Admin Dashboard</h2>
+                    <p>Quản lý và phân phối chat</p>
                 </div>
 
-                <!-- Chat List -->
-                <div id="chatList" class="flex-1 overflow-y-auto custom-scrollbar">
-                    <!-- Chat items will be populated here -->
-                </div>
-            </div>
+                <div class="chat-list">
+                    @forelse($conversations as $conversation)
+                        @php
+                            $unreadCount = $conversation->messages->where('is_read', false)->count();
+                            $firstLetter = strtoupper(
+                                substr(
+                                    $conversation->customer->full_name ?? ($conversation->customer->name ?? 'K'),
+                                    0,
+                                    1,
+                                ),
+                            );
+                        @endphp
+                        <div class="chat-item {{ $loop->first ? 'active' : '' }}"
+                            data-conversation-id="{{ $conversation->id }}"
+                            data-customer-name="{{ $conversation->customer->full_name ?? ($conversation->customer->name ?? 'Khách hàng') }}"
+                            data-customer-email="{{ $conversation->customer->email ?? 'test@customer.com' }}"
+                            data-status="{{ $conversation->status }}">
+                            @if ($unreadCount > 0)
+                                <div class="unread-badge">{{ $unreadCount }}</div>
+                            @endif
+                            <div class="chat-item-header">
+                                <h3 class="chat-item-name">
+                                    {{ $conversation->customer->full_name ?? ($conversation->customer->name ?? 'Khách hàng') }}
+                                </h3>
+                                <span class="chat-item-time">{{ $conversation->updated_at->format('H:i') }}</span>
+                            </div>
+                            <p class="chat-item-email">{{ $conversation->customer->email ?? 'test@customer.com' }}</p>
+                            <p class="chat-item-preview">
+                                @if ($conversation->messages->last())
+                                    {{ Str::limit($conversation->messages->last()->content ?? ($conversation->messages->last()->message ?? 'Tệp đính kèm'), 60) }}
+                                @else
+                                    {{ $conversation->subject ?? 'Cuộc trò chuyện mới' }}
+                                @endif
+                            </p>
+                            <div class="chat-item-footer">
+                                <div class="chat-item-badges">
+                                    @switch($conversation->status)
+                                        @case('new')
+                                            <span class="badge badge-waiting">Chờ xử lý</span>
+                                            <span class="badge badge-high">Cao</span>
+                                        @break
 
-            <!-- Main Chat Area -->
-            <div class="flex-1 flex flex-col">
-                <!-- No Chat Selected State -->
-                <div id="noChatSelected" class="flex-1 flex items-center justify-center bg-muted/30">
-                    <div class="text-center">
-                        <div class="w-24 h-24 bg-muted rounded-full flex items-center justify-center mx-auto mb-4">
-                            <i class="fas fa-comments text-3xl text-muted-foreground"></i>
-                        </div>
-                        <h3 class="text-lg font-medium text-foreground mb-2">Chọn cuộc trò chuyện</h3>
-                        <p class="text-muted-foreground">Chọn một cuộc trò chuyện từ danh sách để bắt đầu hỗ trợ khách hàng
-                        </p>
-                    </div>
-                </div>
+                                        @case('distributed')
+                                            <span class="badge badge-distributed">Đã phân công</span>
+                                        @break
 
-                <!-- Active Chat -->
-                <div id="activeChat" class="flex-1 flex-col hidden">
-                    <!-- Chat Header -->
-                    <div id="chatHeader" class="bg-card border-b border-border px-6 py-4">
-                        <!-- Will be populated when chat is selected -->
-                    </div>
+                                        @case('closed')
+                                            <span class="badge badge-waiting">Đã đóng</span>
+                                        @break
 
-                    <!-- Messages Area -->
-                    <div id="messagesArea" class="flex-1 overflow-y-auto p-6 bg-muted/30 custom-scrollbar">
-                        <!-- Messages will be populated here -->
-                    </div>
+                                        @default
+                                            <span class="badge badge-waiting">Đang xử lý</span>
+                                    @endswitch
 
-                    <!-- Quick Replies -->
-                    <div class="bg-card border-t border-border px-6 py-3">
-                        <div class="flex gap-2 flex-wrap">
-                            <button class="quick-reply-btn px-3 py-1 text-sm rounded-full transition-colors">
-                                Xin chào! Tôi có thể giúp gì cho bạn?
-                            </button>
-                            <button class="quick-reply-btn px-3 py-1 text-sm rounded-full transition-colors">
-                                Để tôi kiểm tra thông tin cho bạn
-                            </button>
-                            <button class="quick-reply-btn px-3 py-1 text-sm rounded-full transition-colors">
-                                Cảm ơn bạn đã liên hệ với FastFood
-                            </button>
-                            <button class="quick-reply-btn px-3 py-1 text-sm rounded-full transition-colors">
-                                Bạn có cần hỗ trợ thêm không?
-                            </button>
-                        </div>
-                    </div>
-
-                    <!-- Message Input -->
-                    <div class="bg-card border-t border-border px-6 py-4">
-                        <div class="flex items-end gap-4">
-                            <div class="flex-1">
-                                <div class="flex items-center gap-2 mb-3">
-                                    <button id="attachFileBtn"
-                                        class="p-2 text-muted-foreground hover:text-primary hover:bg-primary/10 rounded-lg transition-colors">
-                                        <i class="fas fa-paperclip"></i>
-                                    </button>
-                                    <button id="attachImageBtn"
-                                        class="p-2 text-muted-foreground hover:text-primary hover:bg-primary/10 rounded-lg transition-colors">
-                                        <i class="fas fa-image"></i>
-                                    </button>
-                                    <button id="emojiBtn"
-                                        class="p-2 text-muted-foreground hover:text-primary hover:bg-primary/10 rounded-lg transition-colors">
-                                        <i class="fas fa-smile"></i>
-                                    </button>
-                                    <button id="templateBtn"
-                                        class="p-2 text-muted-foreground hover:text-primary hover:bg-primary/10 rounded-lg transition-colors">
-                                        <i class="fas fa-file-alt"></i>
-                                    </button>
-                                </div>
-
-                                <div class="relative">
-                                    <textarea id="messageInput" placeholder="Nhập tin nhắn..."
-                                        class="w-full min-h-[60px] max-h-[120px] resize-none border border-input rounded-lg px-4 py-3 pr-12 bg-background text-foreground focus:ring-2 focus:ring-primary focus:border-primary"
-                                        rows="2"></textarea>
-
-                                    <!-- Emoji Picker -->
-                                    <div id="emojiPicker"
-                                        class="absolute bottom-full right-0 mb-2 bg-popover border border-border rounded-lg p-3 shadow-lg z-10 hidden">
-                                        <div class="grid grid-cols-8 gap-1" id="emojiGrid">
-                                            <!-- Emojis will be populated here -->
-                                        </div>
-                                    </div>
                                 </div>
                             </div>
 
-                            <div class="flex flex-col gap-2">
-                                <button id="sendBtn"
-                                    class="btn btn-primary px-6 py-3 rounded-lg disabled:opacity-50 disabled:cursor-not-allowed"
-                                    disabled>
-                                    <i class="fas fa-paper-plane"></i>
-                                </button>
-                                <button id="closeChatBtn" class="btn btn-destructive px-6 py-3 rounded-lg">
-                                    <i class="fas fa-times"></i>
-                                </button>
+                            @if ($conversation->status === 'new')
+                                <select class="distribution-select" data-conversation-id="{{ $conversation->id }}">
+                                    <option value="" disabled selected>Phân công cho chi nhánh</option>
+                                    @foreach ($branches as $branch)
+                                        <option value="{{ $branch->id }}">{{ $branch->name }}</option>
+                                    @endforeach
+                                </select>
+                            @endif
+                        </div>
+                        @empty
+                            <div class="empty-state">
+                                <i class="fas fa-inbox"></i>
+                                <h3>Không có cuộc trò chuyện</h3>
+                                <p>Chưa có cuộc trò chuyện nào được tạo</p>
+                            </div>
+                        @endforelse
+                    </div>
+                </div>
+
+                <!-- Chat Main Area -->
+                <div class="chat-main" id="chat-main">
+                    @if ($conversations->count() > 0)
+                        @php $firstConversation = $conversations->first(); @endphp
+                        <!-- Chat Header -->
+                        <div class="chat-header" id="chat-header">
+                            <div class="chat-header-user">
+                                <div class="chat-avatar" id="chat-avatar">
+                                    {{ strtoupper(substr($firstConversation->customer->full_name ?? ($firstConversation->customer->name ?? 'K'), 0, 1)) }}
+                                </div>
+                                <div class="chat-header-info">
+                                    <h3 id="chat-customer-name">
+                                        {{ $firstConversation->customer->full_name ?? ($firstConversation->customer->name ?? 'Khách hàng') }}
+                                    </h3>
+                                    <p id="chat-customer-email">
+                                        {{ $firstConversation->customer->email ?? 'test@customer.com' }}
+                                    </p>
+                                </div>
+                            </div>
+                            <div class="chat-header-actions" id="chat-header-actions">
+                                @switch($firstConversation->status)
+                                    @case('new')
+                                        <span class="badge badge-waiting">Chờ xử lý</span>
+                                        <span class="badge badge-high">Cao</span>
+                                    @break
+
+                                    @case('distributed')
+                                        <span class="badge badge-distributed">Đã phân công</span>
+                                    @break
+
+                                    @case('closed')
+                                        <span class="badge badge-waiting">Đã đóng</span>
+                                    @break
+
+                                    @default
+                                        <span class="badge badge-waiting">Đang xử lý</span>
+                                @endswitch
                             </div>
                         </div>
-                    </div>
+
+                        <!-- Chat Messages -->
+                        <div class="chat-messages" id="chat-messages">
+                            <!-- Messages will be loaded here via AJAX -->
+                        </div>
+
+                        <!-- Chat Input -->
+                        <div class="chat-input-container">
+                            <button class="chat-tools-btn" id="attachFileBtn">
+                                <i class="fas fa-paperclip"></i>
+                            </button>
+                            <textarea id="message-input" class="chat-input" placeholder="Nhập tin nhắn..." rows="1"></textarea>
+                            <button id="sendBtn" class="chat-send-btn" disabled>
+                                <i class="fas fa-paper-plane"></i>
+                            </button>
+                        </div>
+                    @else
+                        <div class="empty-state">
+                            <i class="fas fa-comments"></i>
+                            <h3>Chọn cuộc trò chuyện</h3>
+                            <p>Chọn một cuộc trò chuyện từ danh sách để bắt đầu nhắn tin</p>
+                        </div>
+                    @endif
                 </div>
             </div>
 
-            <!-- Customer Info Sidebar -->
-            <div id="customerInfo" class="w-80 bg-card border-l border-border hidden">
-                <div class="p-6">
-                    <h3 class="text-lg font-semibold mb-4 text-foreground">Thông tin khách hàng</h3>
-                    <div id="customerDetails">
-                        <!-- Customer details will be populated here -->
-                    </div>
-                </div>
-            </div>
+            <!-- Hidden file input -->
+            <input type="file" id="fileInput" style="display: none;" accept="image/*,.pdf,.doc,.docx,.txt,.zip,.rar">
         </div>
-    </div>
-
-    <!-- Templates Modal -->
-    <div id="templatesModal" class="fixed inset-0 modal-overlay flex items-center justify-center z-50 hidden">
-        <div class="modal-content rounded-lg w-full max-w-2xl mx-4 max-h-[80vh] overflow-hidden">
-            <div class="flex items-center justify-between p-6 border-b border-border">
-                <h3 class="text-lg font-semibold text-foreground">Mẫu tin nhắn</h3>
-                <button id="closeTemplatesBtn" class="text-muted-foreground hover:text-foreground">
-                    <i class="fas fa-times"></i>
-                </button>
-            </div>
-
-            <div class="p-6 overflow-y-auto max-h-[60vh]">
-                <div class="grid gap-4">
-                    <div
-                        class="template-item p-4 border border-border rounded-lg hover:border-primary cursor-pointer transition-colors">
-                        <h4 class="font-medium mb-2 text-foreground">Chào mừng khách hàng</h4>
-                        <p class="text-sm text-muted-foreground">Xin chào! Cảm ơn bạn đã liên hệ với FastFood. Tôi có thể
-                            giúp gì cho bạn hôm nay?</p>
-                    </div>
-
-                    <div
-                        class="template-item p-4 border border-border rounded-lg hover:border-primary cursor-pointer transition-colors">
-                        <h4 class="font-medium mb-2 text-foreground">Hỏi thông tin đơn hàng</h4>
-                        <p class="text-sm text-muted-foreground">Bạn có thể cung cấp mã đơn hàng để tôi kiểm tra thông tin
-                            cho bạn không?</p>
-                    </div>
-
-                    <div
-                        class="template-item p-4 border border-border rounded-lg hover:border-primary cursor-pointer transition-colors">
-                        <h4 class="font-medium mb-2 text-foreground">Hướng dẫn đặt hàng</h4>
-                        <p class="text-sm text-muted-foreground">Để đặt hàng, bạn có thể truy cập website của chúng tôi
-                            hoặc gọi hotline 1900-xxxx. Chúng tôi sẽ giao hàng trong vòng 30 phút.</p>
-                    </div>
-
-                    <div
-                        class="template-item p-4 border border-border rounded-lg hover:border-primary cursor-pointer transition-colors">
-                        <h4 class="font-medium mb-2 text-foreground">Thông tin khuyến mãi</h4>
-                        <p class="text-sm text-muted-foreground">Hiện tại chúng tôi có chương trình giảm 20% cho đơn hàng
-                            từ 200k và miễn phí giao hàng trong bán kính 5km.</p>
-                    </div>
-
-                    <div
-                        class="template-item p-4 border border-border rounded-lg hover:border-primary cursor-pointer transition-colors">
-                        <h4 class="font-medium mb-2 text-foreground">Hỗ trợ kỹ thuật</h4>
-                        <p class="text-sm text-muted-foreground">Nếu bạn gặp vấn đề với ứng dụng, vui lòng thử tắt và mở
-                            lại app. Nếu vẫn không được, tôi sẽ chuyển cho bộ phận kỹ thuật.</p>
-                    </div>
-
-                    <div
-                        class="template-item p-4 border border-border rounded-lg hover:border-primary cursor-pointer transition-colors">
-                        <h4 class="font-medium mb-2 text-foreground">Kết thúc hỗ trợ</h4>
-                        <p class="text-sm text-muted-foreground">Cảm ơn bạn đã liên hệ với chúng tôi. Nếu có thắc mắc gì
-                            khác, đừng ngần ngại liên hệ nhé! Chúc bạn một ngày tốt lành!</p>
-                    </div>
-                </div>
-            </div>
         </div>
-    </div>
 
-    <!-- Hidden file inputs -->
-    <input type="file" id="fileInput" class="hidden" accept=".pdf,.doc,.docx,.txt,.zip,.rar">
-    <input type="file" id="imageInput" class="hidden" accept="image/*">
-
-@endsection
-
-@section('scripts')
-    <script>
-        document.addEventListener('DOMContentLoaded', function() {
-            // Theme Management
-            const themeToggle = document.getElementById('themeToggle');
-            const themeIcon = document.getElementById('themeIcon');
-            const html = document.documentElement;
-
-            // Load saved theme or default to light
-            const savedTheme = localStorage.getItem('theme') || 'light';
-            setTheme(savedTheme);
-
-            function setTheme(theme) {
-                if (theme === 'dark') {
-                    html.classList.add('dark');
-                    themeToggle.classList.add('dark');
-                    themeIcon.textContent = '☀️';
-                } else {
-                    html.classList.remove('dark');
-                    themeToggle.classList.remove('dark');
-                    themeIcon.textContent = '🌙';
-                }
-                localStorage.setItem('theme', theme);
-            }
-
-            themeToggle.addEventListener('click', function() {
-                const currentTheme = html.classList.contains('dark') ? 'dark' : 'light';
-                const newTheme = currentTheme === 'dark' ? 'light' : 'dark';
-                setTheme(newTheme);
-            });
-
-            // Admin Chat Management System
-            const chatList = document.getElementById('chatList');
-            const noChatSelected = document.getElementById('noChatSelected');
-            const activeChat = document.getElementById('activeChat');
-            const chatHeader = document.getElementById('chatHeader');
-            const messagesArea = document.getElementById('messagesArea');
-            const messageInput = document.getElementById('messageInput');
-            const sendBtn = document.getElementById('sendBtn');
-            const customerInfo = document.getElementById('customerInfo');
-            const customerDetails = document.getElementById('customerDetails');
-            const statusToggle = document.getElementById('statusToggle');
-            const statusText = document.getElementById('statusText');
-            const notificationBadge = document.getElementById('notificationBadge');
-            const searchChats = document.getElementById('searchChats');
-            const filterStatus = document.getElementById('filterStatus');
-            const refreshChats = document.getElementById('refreshChats');
-            const templatesModal = document.getElementById('templatesModal');
-            const templateBtn = document.getElementById('templateBtn');
-            const closeTemplatesBtn = document.getElementById('closeTemplatesBtn');
-            const emojiBtn = document.getElementById('emojiBtn');
-            const emojiPicker = document.getElementById('emojiPicker');
-            const emojiGrid = document.getElementById('emojiGrid');
-            const attachFileBtn = document.getElementById('attachFileBtn');
-            const attachImageBtn = document.getElementById('attachImageBtn');
-            const fileInput = document.getElementById('fileInput');
-            const imageInput = document.getElementById('imageInput');
-            const closeChatBtn = document.getElementById('closeChatBtn');
-
-            // State
-            let isAdminOnline = true;
-            let currentChatId = null;
-            let chats = [];
-            let unreadCount = 0;
-
-            // Emojis
-            const emojis = ['😊', '😂', '❤️', '👍', '👎', '😢', '😮', '😡', '🙏', '👏', '🎉', '🔥', '💯', '✅', '❌',
-                '⚡', '🍕', '🍔', '🍟', '🥤'
-            ];
-
-            // Initialize
-            init();
-
-            function init() {
-                setupEventListeners();
-                populateEmojis();
-                loadChats();
-                startNotificationCheck();
-            }
-
-            function setupEventListeners() {
-                statusToggle.addEventListener('click', toggleAdminStatus);
-                sendBtn.addEventListener('click', sendMessage);
-                messageInput.addEventListener('input', handleInputChange);
-                messageInput.addEventListener('keypress', handleKeyPress);
-                searchChats.addEventListener('input', filterChats);
-                filterStatus.addEventListener('change', filterChats);
-                refreshChats.addEventListener('click', loadChats);
-                templateBtn.addEventListener('click', showTemplatesModal);
-                closeTemplatesBtn.addEventListener('click', hideTemplatesModal);
-                emojiBtn.addEventListener('click', toggleEmojiPicker);
-                attachFileBtn.addEventListener('click', () => fileInput.click());
-                attachImageBtn.addEventListener('click', () => imageInput.click());
-                fileInput.addEventListener('change', handleFileUpload);
-                imageInput.addEventListener('change', handleImageUpload);
-                closeChatBtn.addEventListener('click', closeCurrentChat);
-
-                // Quick reply buttons
-                document.addEventListener('click', function(e) {
-                    if (e.target.classList.contains('quick-reply-btn')) {
-                        messageInput.value = e.target.textContent;
-                        handleInputChange();
-                        messageInput.focus();
-                    }
-                });
-
-                // Template selection
-                document.addEventListener('click', function(e) {
-                    if (e.target.closest('.template-item')) {
-                        const template = e.target.closest('.template-item');
-                        const text = template.querySelector('p').textContent;
-                        messageInput.value = text;
-                        handleInputChange();
-                        hideTemplatesModal();
-                        messageInput.focus();
-                    }
-                });
-
-                // Close emoji picker when clicking outside
-                document.addEventListener('click', function(e) {
-                    if (!emojiPicker.contains(e.target) && !emojiBtn.contains(e.target)) {
-                        hideEmojiPicker();
-                    }
-                });
-
-                // Auto-resize textarea
-                messageInput.addEventListener('input', autoResizeTextarea);
-            }
-
-            function populateEmojis() {
-                emojiGrid.innerHTML = '';
-                emojis.forEach(emoji => {
-                    const button = document.createElement('button');
-                    button.textContent = emoji;
-                    button.className = 'p-2 hover:bg-accent rounded text-lg transition-colors';
-                    button.addEventListener('click', () => addEmoji(emoji));
-                    emojiGrid.appendChild(button);
-                });
-            }
-
-            function toggleAdminStatus() {
-                isAdminOnline = !isAdminOnline;
-                updateStatusDisplay();
-
-                // Send status update to server
-                // axios.post('/api/admin/chat/status', {
-                //     online: isAdminOnline
-                // }).catch(error => {
-                //     console.error('Failed to update status:', error);
-                // });
-            }
-
-            function updateStatusDisplay() {
-                const toggle = statusToggle.querySelector('span');
-
-                if (isAdminOnline) {
-                    statusToggle.className =
-                        'relative inline-flex h-6 w-11 items-center rounded-full bg-green-500 transition-colors';
-                    toggle.className =
-                        'inline-block h-4 w-4 transform rounded-full bg-white transition-transform translate-x-6';
-                    statusText.textContent = 'Online';
-                    statusText.className = 'text-sm font-medium text-green-600';
-                } else {
-                    statusToggle.className =
-                        'relative inline-flex h-6 w-11 items-center rounded-full bg-gray-300 transition-colors';
-                    toggle.className =
-                        'inline-block h-4 w-4 transform rounded-full bg-white transition-transform translate-x-1';
-                    statusText.textContent = 'Offline';
-                    statusText.className = 'text-sm font-medium text-gray-600';
-                }
-            }
-
-            function loadChats() {
-                // Enhanced sample data with more realistic conversations
-                chats = [{
-                        id: '1',
-                        customer: {
-                            name: 'Nguyễn Văn Minh',
-                            email: 'nguyenvanminh@gmail.com',
-                            phone: '0123456789',
-                            avatar: '/placeholder.svg?height=40&width=40'
-                        },
-                        lastMessage: 'Tôi muốn hỏi về combo gia đình mới',
-                        lastMessageTime: new Date(Date.now() - 120000), // 2 minutes ago
-                        status: 'waiting',
-                        unread: 3,
-                        messages: [{
-                                id: '1',
-                                content: 'Xin chào! Tôi muốn hỏi về combo gia đình mới mà FastFood vừa ra mắt',
-                                sender: 'customer',
-                                timestamp: new Date(Date.now() - 300000),
-                                type: 'text'
-                            },
-                            {
-                                id: '2',
-                                content: 'Combo này có những món gì và giá bao nhiêu ạ?',
-                                sender: 'customer',
-                                timestamp: new Date(Date.now() - 240000),
-                                type: 'text'
-                            },
-                            {
-                                id: '3',
-                                content: 'Tôi muốn hỏi về combo gia đình mới',
-                                sender: 'customer',
-                                timestamp: new Date(Date.now() - 120000),
-                                type: 'text'
-                            }
-                        ]
-                    },
-                    {
-                        id: '2',
-                        customer: {
-                            name: 'Trần Thị Hương',
-                            email: 'tranhuong2024@email.com',
-                            phone: '0987654321',
-                            avatar: '/placeholder.svg?height=40&width=40'
-                        },
-                        lastMessage: 'Cảm ơn bạn đã hỗ trợ nhiệt tình!',
-                        lastMessageTime: new Date(Date.now() - 480000), // 8 minutes ago
-                        status: 'active',
-                        unread: 0,
-                        messages: [{
-                                id: '1',
-                                content: 'Chào bạn! Tôi vừa đặt đơn hàng nhưng quên không chọn thêm nước uống',
-                                sender: 'customer',
-                                timestamp: new Date(Date.now() - 1200000),
-                                type: 'text'
-                            },
-                            {
-                                id: '2',
-                                content: 'Xin chào! Để tôi kiểm tra đơn hàng của bạn. Bạn có thể cung cấp mã đơn hàng không?',
-                                sender: 'admin',
-                                timestamp: new Date(Date.now() - 1080000),
-                                type: 'text'
-                            },
-                            {
-                                id: '3',
-                                content: 'Mã đơn hàng là FF2024002 ạ',
-                                sender: 'customer',
-                                timestamp: new Date(Date.now() - 1020000),
-                                type: 'text'
-                            },
-                            {
-                                id: '4',
-                                content: 'Tôi đã cập nhật đơn hàng và thêm 2 ly Coca cho bạn. Không tính thêm phí giao hàng nhé!',
-                                sender: 'admin',
-                                timestamp: new Date(Date.now() - 900000),
-                                type: 'text'
-                            },
-                            {
-                                id: '5',
-                                content: 'Cảm ơn bạn đã hỗ trợ nhiệt tình!',
-                                sender: 'customer',
-                                timestamp: new Date(Date.now() - 480000),
-                                type: 'text'
-                            }
-                        ]
-                    },
-                    {
-                        id: '3',
-                        customer: {
-                            name: 'Lê Hoàng Nam',
-                            email: 'hoangnam.dev@outlook.com',
-                            phone: '0369852147',
-                            avatar: '/placeholder.svg?height=40&width=40'
-                        },
-                        lastMessage: 'Đơn hàng #FF2024001 của tôi bị delay không ạ?',
-                        lastMessageTime: new Date(Date.now() - 900000), // 15 minutes ago
-                        status: 'waiting',
-                        unread: 1,
-                        messages: [{
-                                id: '1',
-                                content: 'Chào admin! Tôi đặt đơn hàng từ 45 phút trước rồi mà chưa thấy shipper liên hệ',
-                                sender: 'customer',
-                                timestamp: new Date(Date.now() - 1200000),
-                                type: 'text'
-                            },
-                            {
-                                id: '2',
-                                content: 'Đơn hàng #FF2024001 của tôi bị delay không ạ?',
-                                sender: 'customer',
-                                timestamp: new Date(Date.now() - 900000),
-                                type: 'text'
-                            }
-                        ]
-                    },
-                    {
-                        id: '4',
-                        customer: {
-                            name: 'Phạm Thị Lan',
-                            email: 'phamlan.work@gmail.com',
-                            phone: '0912345678',
-                            avatar: '/placeholder.svg?height=40&width=40'
-                        },
-                        lastMessage: 'Tôi có thể đổi địa chỉ giao hàng được không?',
-                        lastMessageTime: new Date(Date.now() - 1500000), // 25 minutes ago
-                        status: 'active',
-                        unread: 0,
-                        messages: [{
-                                id: '1',
-                                content: 'Xin chào! Tôi vừa đặt đơn hàng nhưng cần đổi địa chỉ giao hàng',
-                                sender: 'customer',
-                                timestamp: new Date(Date.now() - 1800000),
-                                type: 'text'
-                            },
-                            {
-                                id: '2',
-                                content: 'Chào bạn! Tôi có thể hỗ trợ bạn thay đổi địa chỉ. Bạn cho tôi mã đơn hàng nhé',
-                                sender: 'admin',
-                                timestamp: new Date(Date.now() - 1680000),
-                                type: 'text'
-                            },
-                            {
-                                id: '3',
-                                content: 'Mã đơn hàng FF2024003. Địa chỉ mới là: 123 Nguyễn Văn Linh, Quận 7, TP.HCM',
-                                sender: 'customer',
-                                timestamp: new Date(Date.now() - 1620000),
-                                type: 'text'
-                            },
-                            {
-                                id: '4',
-                                content: 'Đã cập nhật địa chỉ mới cho bạn. Shipper sẽ giao đến địa chỉ mới trong 20 phút nữa',
-                                sender: 'admin',
-                                timestamp: new Date(Date.now() - 1560000),
-                                type: 'text'
-                            },
-                            {
-                                id: '5',
-                                content: 'Tôi có thể đổi địa chỉ giao hàng được không?',
-                                sender: 'customer',
-                                timestamp: new Date(Date.now() - 1500000),
-                                type: 'text'
-                            }
-                        ]
-                    },
-                    {
-                        id: '5',
-                        customer: {
-                            name: 'Võ Minh Tuấn',
-                            email: 'vominhtuan88@yahoo.com',
-                            phone: '0834567890',
-                            avatar: '/placeholder.svg?height=40&width=40'
-                        },
-                        lastMessage: 'Món ăn rất ngon, cảm ơn FastFood!',
-                        lastMessageTime: new Date(Date.now() - 3600000), // 1 hour ago
-                        status: 'closed',
-                        unread: 0,
-                        messages: [{
-                                id: '1',
-                                content: 'Tôi vừa nhận được đơn hàng. Món ăn rất ngon!',
-                                sender: 'customer',
-                                timestamp: new Date(Date.now() - 3720000),
-                                type: 'text'
-                            },
-                            {
-                                id: '2',
-                                content: 'Cảm ơn bạn đã đánh giá! Chúng tôi rất vui khi bạn hài lòng với dịch vụ',
-                                sender: 'admin',
-                                timestamp: new Date(Date.now() - 3660000),
-                                type: 'text'
-                            },
-                            {
-                                id: '3',
-                                content: 'Món ăn rất ngon, cảm ơn FastFood!',
-                                sender: 'customer',
-                                timestamp: new Date(Date.now() - 3600000),
-                                type: 'text'
-                            }
-                        ]
-                    },
-                    {
-                        id: '6',
-                        customer: {
-                            name: 'Đặng Thị Mai',
-                            email: 'dangmai.student@edu.vn',
-                            phone: '0756789012',
-                            avatar: '/placeholder.svg?height=40&width=40'
-                        },
-                        lastMessage: 'Có chương trình khuyến mãi nào cho sinh viên không ạ?',
-                        lastMessageTime: new Date(Date.now() - 7200000), // 2 hours ago
-                        status: 'waiting',
-                        unread: 2,
-                        messages: [{
-                                id: '1',
-                                content: 'Chào FastFood! Mình là sinh viên, có chương trình ưu đãi gì không ạ?',
-                                sender: 'customer',
-                                timestamp: new Date(Date.now() - 7500000),
-                                type: 'text'
-                            },
-                            {
-                                id: '2',
-                                content: 'Có chương trình khuyến mãi nào cho sinh viên không ạ?',
-                                sender: 'customer',
-                                timestamp: new Date(Date.now() - 7200000),
-                                type: 'text'
-                            }
-                        ]
-                    },
-                    {
-                        id: '7',
-                        customer: {
-                            name: 'Bùi Văn Đức',
-                            email: 'buivanduc.biz@company.com',
-                            phone: '0678901234',
-                            avatar: '/placeholder.svg?height=40&width=40'
-                        },
-                        lastMessage: 'Tôi muốn đặt tiệc cho 50 người',
-                        lastMessageTime: new Date(Date.now() - 10800000), // 3 hours ago
-                        status: 'active',
-                        unread: 0,
-                        messages: [{
-                                id: '1',
-                                content: 'Xin chào! Công ty tôi muốn đặt tiệc cho sự kiện 50 người',
-                                sender: 'customer',
-                                timestamp: new Date(Date.now() - 11400000),
-                                type: 'text'
-                            },
-                            {
-                                id: '2',
-                                content: 'Chào anh! Chúng tôi có gói tiệc doanh nghiệp rất phù hợp. Tôi sẽ gửi bảng giá cho anh',
-                                sender: 'admin',
-                                timestamp: new Date(Date.now() - 11280000),
-                                type: 'text'
-                            },
-                            {
-                                id: '3',
-                                content: 'Tôi muốn đặt tiệc cho 50 người',
-                                sender: 'customer',
-                                timestamp: new Date(Date.now() - 10800000),
-                                type: 'text'
-                            }
-                        ]
-                    },
-                    {
-                        id: '8',
-                        customer: {
-                            name: 'Hoàng Thị Linh',
-                            email: 'hoanglinh.designer@creative.vn',
-                            phone: '0590123456',
-                            avatar: '/placeholder.svg?height=40&width=40'
-                        },
-                        lastMessage: 'Cảm ơn đã giải quyết vấn đề!',
-                        lastMessageTime: new Date(Date.now() - 18000000), // 5 hours ago
-                        status: 'closed',
-                        unread: 0,
-                        messages: [{
-                                id: '1',
-                                content: 'Tôi gặp lỗi khi thanh toán qua ví điện tử',
-                                sender: 'customer',
-                                timestamp: new Date(Date.now() - 18600000),
-                                type: 'text'
-                            },
-                            {
-                                id: '2',
-                                content: 'Tôi sẽ kiểm tra và hỗ trợ bạn ngay. Bạn đang sử dụng ví nào?',
-                                sender: 'admin',
-                                timestamp: new Date(Date.now() - 18480000),
-                                type: 'text'
-                            },
-                            {
-                                id: '3',
-                                content: 'Đã khắc phục xong lỗi thanh toán. Bạn thử lại nhé!',
-                                sender: 'admin',
-                                timestamp: new Date(Date.now() - 18120000),
-                                type: 'text'
-                            },
-                            {
-                                id: '4',
-                                content: 'Cảm ơn đã giải quyết vấn đề!',
-                                sender: 'customer',
-                                timestamp: new Date(Date.now() - 18000000),
-                                type: 'text'
-                            }
-                        ]
-                    }
-                ];
-
-                renderChatList();
-                updateNotificationBadge();
-            }
-
-            function renderChatList() {
-                chatList.innerHTML = '';
-
-                const filteredChats = getFilteredChats();
-
-                filteredChats.forEach(chat => {
-                    const chatItem = createChatListItem(chat);
-                    chatList.appendChild(chatItem);
-                });
-            }
-
-            function getFilteredChats() {
-                let filtered = chats;
-
-                // Filter by search
-                const searchTerm = searchChats.value.toLowerCase();
-                if (searchTerm) {
-                    filtered = filtered.filter(chat =>
-                        chat.customer.name.toLowerCase().includes(searchTerm) ||
-                        chat.lastMessage.toLowerCase().includes(searchTerm)
-                    );
-                }
-
-                // Filter by status
-                const statusFilter = filterStatus.value;
-                if (statusFilter !== 'all') {
-                    filtered = filtered.filter(chat => chat.status === statusFilter);
-                }
-
-                return filtered.sort((a, b) => b.lastMessageTime - a.lastMessageTime);
-            }
-
-            function createChatListItem(chat) {
-                const div = document.createElement('div');
-                div.className =
-                    `chat-item p-4 border-b border-border hover:bg-accent cursor-pointer transition-colors ${currentChatId === chat.id ? 'active' : ''}`;
-                div.dataset.chatId = chat.id;
-
-                const timeString = formatTime(chat.lastMessageTime);
-                const statusColor = getStatusColor(chat.status);
-
-                div.innerHTML = `
-            <div class="flex items-start gap-3">
-                <div class="relative">
-                    <img src="${chat.customer.avatar}" alt="${chat.customer.name}" class="w-12 h-12 rounded-full object-cover">
-                    <div class="absolute -bottom-1 -right-1 w-4 h-4 ${statusColor} rounded-full border-2 border-card"></div>
-                </div>
-                
-                <div class="flex-1 min-w-0">
-                    <div class="flex items-center justify-between mb-1">
-                        <h4 class="font-medium text-foreground truncate">${chat.customer.name}</h4>
-                        <span class="text-xs text-muted-foreground">${timeString}</span>
-                    </div>
-                    
-                    <p class="text-sm text-muted-foreground truncate mb-1">${chat.lastMessage}</p>
-                    
-                    <div class="flex items-center justify-between">
-                        <span class="text-xs px-2 py-1 rounded-full ${getStatusBadgeClass(chat.status)}">${getStatusText(chat.status)}</span>
-                        ${chat.unread > 0 ? `<span class="bg-red-500 text-white text-xs px-2 py-1 rounded-full">${chat.unread}</span>` : ''}
-                    </div>
-                </div>
-            </div>
-        `;
-
-                div.addEventListener('click', () => selectChat(chat.id));
-
-                return div;
-            }
-
-            function selectChat(chatId) {
-                currentChatId = chatId;
-                const chat = chats.find(c => c.id === chatId);
-
-                if (chat) {
-                    // Mark as read
-                    chat.unread = 0;
-
-                    // Update UI
-                    renderChatList();
-                    showActiveChat(chat);
-                    updateNotificationBadge();
-                }
-            }
-
-            function showActiveChat(chat) {
-                noChatSelected.classList.add('hidden');
-                activeChat.classList.remove('hidden');
-                activeChat.classList.add('flex');
-                customerInfo.classList.remove('hidden');
-
-                // Update chat header
-                chatHeader.innerHTML = `
-            <div class="flex items-center justify-between">
-                <div class="flex items-center gap-3">
-                    <img src="${chat.customer.avatar}" alt="${chat.customer.name}" class="w-10 h-10 rounded-full object-cover">
-                    <div>
-                        <h3 class="font-semibold text-foreground">${chat.customer.name}</h3>
-                        <p class="text-sm text-muted-foreground">${chat.customer.email}</p>
-                    </div>
-                </div>
-                
-                <div class="flex items-center gap-2">
-                    <span class="text-xs px-3 py-1 rounded-full ${getStatusBadgeClass(chat.status)}">${getStatusText(chat.status)}</span>
-                    <button class="p-2 text-muted-foreground hover:text-foreground hover:bg-accent rounded-lg">
-                        <i class="fas fa-phone"></i>
-                    </button>
-                    <button class="p-2 text-muted-foreground hover:text-foreground hover:bg-accent rounded-lg">
-                        <i class="fas fa-video"></i>
-                    </button>
-                    <button class="p-2 text-muted-foreground hover:text-foreground hover:bg-accent rounded-lg">
-                        <i class="fas fa-ellipsis-v"></i>
-                    </button>
-                </div>
-            </div>
-        `;
-
-                // Update customer info
-                customerDetails.innerHTML = `
-            <div class="space-y-4">
-                <div class="text-center">
-                    <img src="${chat.customer.avatar}" alt="${chat.customer.name}" class="w-20 h-20 rounded-full object-cover mx-auto mb-3">
-                    <h4 class="font-semibold text-foreground">${chat.customer.name}</h4>
-                    <p class="text-sm text-muted-foreground">${chat.customer.email}</p>
-                </div>
-                
-                <div class="space-y-3">
-                    <div>
-                        <label class="text-sm font-medium text-foreground">Số điện thoại</label>
-                        <p class="text-sm text-muted-foreground">${chat.customer.phone}</p>
-                    </div>
-                    
-                    <div>
-                        <label class="text-sm font-medium text-foreground">Trạng thái</label>
-                        <p class="text-sm"><span class="px-2 py-1 rounded-full ${getStatusBadgeClass(chat.status)}">${getStatusText(chat.status)}</span></p>
-                    </div>
-                    
-                    <div>
-                        <label class="text-sm font-medium text-foreground">Lần cuối hoạt động</label>
-                        <p class="text-sm text-muted-foreground">${formatTime(chat.lastMessageTime)}</p>
-                    </div>
-                    
-                    <div>
-                        <label class="text-sm font-medium text-foreground">Tổng đơn hàng</label>
-                        <p class="text-sm text-muted-foreground">${Math.floor(Math.random() * 20) + 5} đơn</p>
-                    </div>
-                </div>
-                
-                <div class="pt-4 border-t border-border space-y-2">
-                    <button class="w-full btn btn-primary py-2 px-4 rounded-lg">
-                        Xem lịch sử đơn hàng
-                    </button>
-                    <button class="w-full btn btn-secondary py-2 px-4 rounded-lg">
-                        Ghi chú khách hàng
-                    </button>
-                </div>
-            </div>
-        `;
-
-                // Render messages
-                renderMessages(chat.messages);
-            }
-
-            function renderMessages(messages) {
-                messagesArea.innerHTML = '';
-
-                messages.forEach(message => {
-                    const messageElement = createMessageElement(message);
-                    messagesArea.appendChild(messageElement);
-                });
-
-                scrollToBottom();
-            }
-
-            function createMessageElement(message) {
-                const div = document.createElement('div');
-                div.className = `flex ${message.sender === 'admin' ? 'justify-end' : 'justify-start'} mb-4`;
-
-                const timeString = formatTime(message.timestamp);
-                const isAdmin = message.sender === 'admin';
-
-                let messageContent = '';
-
-                if (message.type === 'text') {
-                    messageContent = `<p class="text-sm whitespace-pre-wrap">${escapeHtml(message.content)}</p>`;
-                } else if (message.type === 'image') {
-                    messageContent = `
-                <div class="space-y-2">
-                    <img src="${message.imageUrl}" alt="Image" class="max-w-full h-auto rounded-lg max-h-64 object-cover">
-                    ${message.content ? `<p class="text-sm">${escapeHtml(message.content)}</p>` : ''}
-                </div>
-            `;
-                } else if (message.type === 'file') {
-                    messageContent = `
-                <div class="flex items-center gap-3 p-3 bg-white/10 rounded-lg min-w-[200px]">
-                    <i class="fas fa-file text-lg"></i>
-                    <div class="flex-1">
-                        <p class="text-sm font-medium">${escapeHtml(message.fileName)}</p>
-                        <p class="text-xs opacity-75">${message.fileSize}</p>
-                    </div>
-                    <button class="p-1 hover:bg-white/20 rounded">
-                        <i class="fas fa-download"></i>
-                    </button>
-                </div>
-            `;
-                }
-
-                div.innerHTML = `
-            <div class="flex gap-3 max-w-[70%] ${isAdmin ? 'flex-row-reverse' : 'flex-row'}">
-                <div class="w-8 h-8 ${isAdmin ? 'bg-primary' : 'bg-blue-500'} rounded-full flex items-center justify-center flex-shrink-0">
-                    <span class="text-white text-xs font-bold">${isAdmin ? 'A' : 'C'}</span>
-                </div>
-                
-                <div class="flex flex-col ${isAdmin ? 'items-end' : 'items-start'}">
-                    <div class="rounded-2xl px-4 py-3 shadow-sm ${
-                        isAdmin 
-                            ? 'message-admin rounded-br-md' 
-                            : 'message-customer rounded-bl-md'
-                    }">
-                        ${messageContent}
-                    </div>
-                    <span class="text-xs text-muted-foreground mt-1 px-2">${timeString}</span>
-                </div>
-            </div>
-        `;
-
-                return div;
-            }
-
-            function sendMessage() {
-                const content = messageInput.value.trim();
-                if (!content || !currentChatId) return;
-
-                const message = {
-                    id: Date.now().toString(),
-                    content: content,
-                    sender: 'admin',
-                    timestamp: new Date(),
-                    type: 'text'
-                };
-
-                // Add to current chat
-                const chat = chats.find(c => c.id === currentChatId);
-                if (chat) {
-                    chat.messages.push(message);
-                    chat.lastMessage = content;
-                    chat.lastMessageTime = new Date();
-                    chat.status = 'active';
-
-                    // Update UI
-                    renderMessages(chat.messages);
-                    renderChatList();
-                }
-
-                // Clear input
-                messageInput.value = '';
-                handleInputChange();
-                autoResizeTextarea();
-
-                // Send to server
-                // axios.post('/api/admin/chat/send', {
-                //     chatId: currentChatId,
-                //     message: content,
-                //     type: 'text'
-                // }).catch(error => {
-                //     console.error('Failed to send message:', error);
-                // });
-            }
-
-            function handleInputChange() {
-                const hasText = messageInput.value.trim().length > 0;
-                sendBtn.disabled = !hasText;
-            }
-
-            function handleKeyPress(e) {
-                if (e.key === 'Enter' && !e.shiftKey) {
-                    e.preventDefault();
-                    sendMessage();
-                }
-            }
-
-            function autoResizeTextarea() {
-                messageInput.style.height = 'auto';
-                messageInput.style.height = Math.min(messageInput.scrollHeight, 120) + 'px';
-            }
-
-            function handleFileUpload(e) {
-                const file = e.target.files[0];
-                if (!file || !currentChatId) return;
-
-                const message = {
-                    id: Date.now().toString(),
-                    content: `Đã gửi file: ${file.name}`,
-                    sender: 'admin',
-                    timestamp: new Date(),
-                    type: 'file',
-                    fileName: file.name,
-                    fileSize: (file.size / 1024 / 1024).toFixed(2) + ' MB',
-                    fileUrl: URL.createObjectURL(file)
-                };
-
-                addMessageToCurrentChat(message);
-                e.target.value = '';
-            }
-
-            function handleImageUpload(e) {
-                const file = e.target.files[0];
-                if (!file || !currentChatId) return;
-
-                const reader = new FileReader();
-                reader.onload = function(e) {
-                    const message = {
-                        id: Date.now().toString(),
-                        content: '',
-                        sender: 'admin',
-                        timestamp: new Date(),
-                        type: 'image',
-                        imageUrl: e.target.result
-                    };
-
-                    addMessageToCurrentChat(message);
-                };
-                reader.readAsDataURL(file);
-
-                e.target.value = '';
-            }
-
-            function addMessageToCurrentChat(message) {
-                const chat = chats.find(c => c.id === currentChatId);
-                if (chat) {
-                    chat.messages.push(message);
-                    chat.lastMessage = message.content || 'Đã gửi file';
-                    chat.lastMessageTime = new Date();
-
-                    renderMessages(chat.messages);
-                    renderChatList();
-                }
-            }
-
-            function closeCurrentChat() {
-                if (!currentChatId) return;
-
-                const chat = chats.find(c => c.id === currentChatId);
-                if (chat) {
-                    chat.status = 'closed';
-                    renderChatList();
-                }
-
-                // Hide active chat
-                activeChat.classList.add('hidden');
-                customerInfo.classList.add('hidden');
-                noChatSelected.classList.remove('hidden');
-                currentChatId = null;
-            }
-
-            function showTemplatesModal() {
-                templatesModal.classList.remove('hidden');
-            }
-
-            function hideTemplatesModal() {
-                templatesModal.classList.add('hidden');
-            }
-
-            function toggleEmojiPicker() {
-                emojiPicker.classList.toggle('hidden');
-            }
-
-            function hideEmojiPicker() {
-                emojiPicker.classList.add('hidden');
-            }
-
-            function addEmoji(emoji) {
-                messageInput.value += emoji;
-                handleInputChange();
-                hideEmojiPicker();
-                messageInput.focus();
-            }
-
-            function filterChats() {
-                renderChatList();
-            }
-
-            function updateNotificationBadge() {
-                unreadCount = chats.reduce((total, chat) => total + chat.unread, 0);
-
-                if (unreadCount > 0) {
-                    notificationBadge.textContent = unreadCount;
-                    notificationBadge.classList.remove('hidden');
-                } else {
-                    notificationBadge.classList.add('hidden');
-                }
-            }
-
-            function startNotificationCheck() {
-                // Check for new messages every 30 seconds
-                setInterval(() => {
-                    // In real app, this would fetch from server
-                    // checkForNewMessages();
-                }, 30000);
-            }
-
-            function scrollToBottom() {
-                setTimeout(() => {
-                    messagesArea.scrollTop = messagesArea.scrollHeight;
-                }, 100);
-            }
-
-            // Utility functions
-            function formatTime(date) {
-                const now = new Date();
-                const diff = now - date;
-
-                if (diff < 60000) { // Less than 1 minute
-                    return 'Vừa xong';
-                } else if (diff < 3600000) { // Less than 1 hour
-                    return Math.floor(diff / 60000) + ' phút trước';
-                } else if (diff < 86400000) { // Less than 1 day
-                    return date.toLocaleTimeString('vi-VN', {
-                        hour: '2-digit',
-                        minute: '2-digit'
-                    });
-                } else {
-                    return date.toLocaleDateString('vi-VN');
-                }
-            }
-
-            function getStatusColor(status) {
-                switch (status) {
-                    case 'active':
-                        return 'bg-green-500';
-                    case 'waiting':
-                        return 'bg-yellow-500';
-                    case 'closed':
-                        return 'bg-gray-400';
-                    default:
-                        return 'bg-gray-400';
-                }
-            }
-
-            function getStatusBadgeClass(status) {
-                switch (status) {
-                    case 'active':
-                        return 'bg-green-100 text-green-800 dark:bg-green-900 dark:text-green-300';
-                    case 'waiting':
-                        return 'bg-yellow-100 text-yellow-800 dark:bg-yellow-900 dark:text-yellow-300';
-                    case 'closed':
-                        return 'bg-gray-100 text-gray-800 dark:bg-gray-800 dark:text-gray-300';
-                    default:
-                        return 'bg-gray-100 text-gray-800 dark:bg-gray-800 dark:text-gray-300';
-                }
-            }
-
-            function getStatusText(status) {
-                switch (status) {
-                    case 'active':
-                        return 'Đang hoạt động';
-                    case 'waiting':
-                        return 'Chờ phản hồi';
-                    case 'closed':
-                        return 'Đã đóng';
-                    default:
-                        return 'Không xác định';
-                }
-            }
-
-            function escapeHtml(text) {
-                const div = document.createElement('div');
-                div.textContent = text;
-                return div.innerHTML;
-            }
-        });
-    </script>
-@endsection
+        <script src="https://js.pusher.com/8.2.0/pusher.min.js"></script>
+        <script src="{{ asset('js/echo.js') }}"></script>
+        <script src="{{ asset('js/chat-common.js') }}" defer></script>
+        <script>
+            document.addEventListener('DOMContentLoaded', function() {
+                        console.log('🚀 Admin Chat Dashboard Loading...');
+
+                        // Global variables
+                        const currentUserId = 1;
+                        let selectedConversationId = {{ $conversations->first()->id ?? 'null' }};
+                        let currentConversationData = null;
+
+                        // DOM elements
+                        const chatMessages = document.getElementById('chat-messages');
+                        const messageInput = document.getElementById('messageInput');
+                        const sendBtn = document.getElementById('sendBtn');
+                        const attachFileBtn = document.getElementById('attachFileBtn');
+                        const fileInput = document.getElementById('fileInput');
+                        const chatHeader = document.getElementById('chat-header');
+                        const chatAvatar = document.getElementById('chat-avatar');
+                        const chatCustomerName = document.getElementById('chat-customer-name');
+                        const chatCustomerEmail = document.getElementById('chat-customer-email');
+                        const chatHeaderActions = document.getElementById('chat-header-actions');
+
+                        console.log('✅ Initialized with conversation ID:', selectedConversationId);
+
+                        // Initialize
+                        init();
+
+                        function init() {
+                            setupEventListeners();
+                            if (selectedConversationId) {
+                                loadConversation(selectedConversationId);
+                                var chatContainer = document.getElementById('chat-container');
+                                if (!chatContainer) return;
+                                var chat = new ChatCommon({
+                                    conversationId: chatContainer.getAttribute('data-conversation-id'),
+                                    userId: chatContainer.getAttribute('data-user-id'),
+                                    userType: chatContainer.getAttribute('data-user-type'),
+                                    api: {
+                                        send: '{{ route('admin.chat.send') }}',
+                                        getMessages: '{{ route('admin.chat.messages', [':id']) }}',
+                                        distribute: '{{ route('admin.chat.distribute') }}'
+                                    }
+                                });
+
+                                // Lắng nghe sự kiện change của select phân công (distribution-select) và gọi chat.distributeConversation(convId, branchId)
+                                var distributeSelects = document.querySelectorAll('.distribution-select');
+                                distributeSelects.forEach(function(select) {
+                                    select.addEventListener('change', function() {
+                                        var convId = this.getAttribute('data-conversation-id');
+                                        var branchId = this.value;
+                                        if (convId && branchId) {
+                                            chat.distributeConversation(convId, branchId);
+                                        }
+                                    });
+                                });
+                            });
+        </script>
+    @endsection
