+++ resources/views/branch/chat.blade.php
@@ -37,6 +37,14 @@
             line-height: 1.6;
         }
 
+        /* Cải thiện giao diện cho tên chi nhánh */
+        .branch-name h4 {
+            font-size: 1.25rem;
+            font-weight: 700;
+            color: #3b82f6;
+            margin-top: 0.5rem;
+        }
+
         .branch-chat-container {
             max-width: 1400px;
             margin: 0 auto;
@@ -872,1022 +880,327 @@
 
     <meta name="csrf-token" content="{{ csrf_token() }}">
 
-    <div class="branch-chat-container">
-        <!-- Enhanced Branch Header -->
-        <div class="branch-header">
-
-            <div class="header-content">
-                <div class="user-info">
-                    <div class="user-avatar">
-                        {{ strtoupper(substr($user->name, 0, 1)) }}
-                    </div>
-                    <div class="user-details">
-                        <h3>{{ $user->name }}</h3>
-                        <div class="role-badge">{{ ucfirst(str_replace('_', ' ', $user->role)) }}</div>
-                    </div>
-                </div>
-                <div class="status-info">
-                    <div class="status-dot"></div>
-                    <div class="status-text">
-                        <strong>Online</strong> • {{ now()->format('H:i') }}
-                    </div>
-                </div>
-            </div>
-        </div>
-
-        <!-- Enhanced Stats Cards -->
-        <div class="stats-container">
-            <div class="stat-card new" onclick="filterConversations('distributed')">
-                <div class="stat-header">
-                    <div class="stat-icon">
-                        <i class="fas fa-inbox"></i>
-                    </div>
-                </div>
-                <div class="stat-number">{{ $conversationsByBranch->where('status', 'distributed')->count() }}</div>
-                <div class="stat-label">Mới nhận</div>
-                <div class="stat-trend">
-                    +{{ $conversationsByBranch->where('status', 'distributed')->where('created_at', '>=', now()->subDay())->count() }}
-                    hôm nay</div>
-            </div>
-
-            <div class="stat-card active" onclick="filterConversations('active')">
-                <div class="stat-header">
-                    <div class="stat-icon">
-                        <i class="fas fa-comments"></i>
-                    </div>
-                </div>
-                <div class="stat-number">{{ $conversations->whereIn('status', ['active', 'open'])->count() }}</div>
-                <div class="stat-label">Đang xử lý</div>
-                <div class="stat-trend">
-                    {{ $conversations->whereIn('status', ['active', 'open'])->where('updated_at', '>=', now()->subHour())->count() }}
-                    hoạt động gần đây</div>
-            </div>
-
-            <div class="stat-card priority" onclick="filterConversations('priority')">
-                <div class="stat-header">
-                    <div class="stat-icon">
-                        <i class="fas fa-exclamation-triangle"></i>
-                    </div>
-                </div>
-                <div class="stat-number">
-                    {{ $conversations->whereIn('priority', ['high', 'urgent'])->count() }}
-                </div>
-                <div class="stat-label">Ưu tiên cao</div>
-                <div class="stat-trend">Cần xử lý ngay</div>
-            </div>
-
-            <div class="stat-card total" onclick="filterConversations('all')">
-                <div class="stat-header">
-                    <div class="stat-icon">
-                        <i class="fas fa-chart-line"></i>
-                    </div>
-                </div>
-                <div class="stat-number">{{ $conversations->count() }}</div>
-                <div class="stat-label">Tổng cộng</div>
-                <div class="stat-trend">{{ $conversations->where('status', 'resolved')->count() }} đã giải quyết</div>
-            </div>
-        </div>
-
-        <!-- Enhanced Chat Interface -->
-        <div class="chat-interface">
-            <!-- Enhanced Sidebar -->
-            <div class="chat-sidebar">
-                <div class="chat-sidebar-header">
-                    <h3>📋 Danh sách cuộc trò chuyện</h3>
-                    <div class="search-container">
-                        <i class="fas fa-search"></i>
-                        <input type="text" id="search-conversations" placeholder="Tìm kiếm khách hàng...">
-                    </div>
-                </div>
-                <div class="conversations-list" id="conversations-list">
-                    @forelse($conversations as $conversation)
-                        @php
-                            $unreadCount = $conversation->messages
-                                ->where('is_read', false)
-                                ->where('sender_id', '!=', $user->id)
-                                ->count();
-                            $customer = $conversation->customer;
-                            $lastMessage = $conversation->messages->last();
-                        @endphp
-                        <div class="conversation-item {{ $loop->first ? 'active' : '' }}"
-                            data-conversation-id="{{ $conversation->id }}" data-status="{{ $conversation->status }}"
-                            data-priority="{{ $conversation->priority ?? 'normal' }}"
-                            data-customer-name="{{ $customer->name ?? 'Khách hàng' }}">
-
-                            <div class="conversation-header">
-                                <div class="customer-name">
-                                    {{ $customer->name ?? 'Khách hàng' }}
-                                </div>
-                                <div class="time">{{ $conversation->updated_at->format('H:i') }}</div>
-                            </div>
-
-                            <div class="message-preview">
-                                @if ($lastMessage)
-                                    @if ($lastMessage->message)
-                                        {{ Str::limit($lastMessage->message, 45) }}
-                                    @elseif ($lastMessage->attachment)
-                                        📎 {{ $lastMessage->attachment_type === 'image' ? 'Hình ảnh' : 'Tệp đính kèm' }}
-                                    @else
-                                        Tin nhắn mới
-                                    @endif
-                                @else
-                                    💬 Cuộc trò chuyện mới
-                                @endif
-                            </div>
-
-                            <div class="conversation-meta">
-                                <span class="status-badge {{ $conversation->status }}">
-                                    @switch($conversation->status)
-                                        @case('distributed')
-                                            📥 Mới nhận
-                                        @break
-
-                                        @case('active')
-                                        @case('open')
-                                            🟢 Đang xử lý
-                                        @break
-
-                                        @case('resolved')
-                                            ✅ Đã giải quyết
-                                        @break
-
-                                        @case('closed')
-                                            🔒 Đã đóng
-                                        @break
-
-                                        @default
-                                            📋 Chờ xử lý
-                                    @endswitch
-                                </span>
-
-                                @if ($unreadCount > 0)
-                                    <div class="unread-badge">{{ $unreadCount }}</div>
-                                @endif
-                            </div>
+    <div id="chat-container" data-conversation-id="{{ $conversations->first()->id ?? '' }}" data-user-id="{{ $user->id }}"
+        data-user-type="branch">
+        <div id="chat-messages" class="chat-messages"></div>
+        <div class="branch-chat-container">
+            <!-- Enhanced Branch Header -->
+            <div class="branch-header">
+                <div class="header-content">
+                    <div class="user-info">
+                        <div class="user-avatar">
+                            {{ strtoupper(substr($user->full_name, 0, 1)) }}
                         </div>
-                        @empty
-                            <div class="empty-state">
-                                <i class="fas fa-inbox"></i>
-                                <h5>Không có cuộc trò chuyện</h5>
-                                <p>Chưa có cuộc trò chuyện nào được phân phối cho chi nhánh này</p>
-                            </div>
-                        @endforelse
+                        <div class="user-details">
+                            <h3>{{ $user->full_name }}</h3>
+                            <div class="role-badge">{{ ucfirst(str_replace('_', ' ', $user->role)) }}</div>
+                        </div>
+                    </div>
+                    <div class="status-info">
+                        <div class="status-dot"></div>
+                        <div class="status-text">
+                            <strong>Online</strong> • {{ now()->format('H:i') }}
+                        </div>
                     </div>
                 </div>
+                {{-- <div class="branch-name">
+                    <h4>{{ $branch->name }}</h4> <!-- Hiển thị tên chi nhánh -->
+                </div> --}}
+            </div>
 
-                <!-- Enhanced Main Chat Area -->
-                <div class="chat-main">
-                    @if ($conversations->count() > 0)
-                        @php $firstConversation = $conversations->first(); @endphp
-                        <!-- Enhanced Chat Header -->
-                        <div class="chat-header">
-                            <div class="customer-info">
-                                <div class="customer-avatar">
-                                    {{ strtoupper(substr($firstConversation->customer->name ?? 'K', 0, 1)) }}
-                                </div>
-                                <div class="customer-details">
-                                    <h2>{{ $firstConversation->customer->name ?? 'Khách hàng' }}</h2>
-                                    <div class="email">{{ $firstConversation->customer->email ?? 'Chưa có email' }}</div>
-                                </div>
-                            </div>
-                            <div class="chat-actions">
-                                <div class="dropdown">
-                                    <button class="btn btn-primary dropdown-toggle" type="button" data-bs-toggle="dropdown"
-                                        aria-expanded="false">
-                                        <i class="fas fa-cog me-2"></i>Hành động
-                                    </button>
-                                    <ul class="dropdown-menu dropdown-menu-end">
-                                        <li><a class="dropdown-item" href="#" onclick="updateStatus('active')">
-                                                <i class="fas fa-play me-2 text-success"></i>Kích hoạt cuộc trò chuyện
-                                            </a></li>
-                                        <li><a class="dropdown-item" href="#" onclick="updateStatus('resolved')">
-                                                <i class="fas fa-check me-2 text-warning"></i>Đánh dấu đã giải quyết
-                                            </a></li>
-                                        <li><a class="dropdown-item" href="#" onclick="updateStatus('closed')">
-                                                <i class="fas fa-times me-2 text-danger"></i>Đóng cuộc trò chuyện
-                                            </a></li>
-                                    </ul>
-                                </div>
-                            </div>
+            <!-- Enhanced Stats Cards -->
+            <div class="stats-container">
+                <div class="stat-card new" onclick="filterConversations('distributed')">
+                    <div class="stat-header">
+                        <div class="stat-icon">
+                            <i class="fas fa-inbox"></i>
                         </div>
+                    </div>
+                    <div class="stat-number">{{ $conversations->where('status', 'distributed')->count() }}</div>
+                    <div class="stat-label">Mới nhận</div>
+                    <div class="stat-trend">
+                        +{{ $conversations->where('status', 'distributed')->where('created_at', '>=', now()->subDay())->count() }}
+                        hôm nay</div>
+                </div>
 
-                        <!-- Enhanced Chat Messages -->
-                        <div class="chat-messages" id="chat-messages">
-                            @if ($firstConversation && $firstConversation->messages->count() > 0)
-                                @foreach ($firstConversation->messages as $message)
-                                    @php
-                                        $isSent = $message->sender_id === $user->id;
-                                        $sender = $message->sender;
-                                        $isSystem = $message->is_system_message ?? false;
-                                    @endphp
+                <div class="stat-card active" onclick="filterConversations('active')">
+                    <div class="stat-header">
+                        <div class="stat-icon">
+                            <i class="fas fa-comments"></i>
+                        </div>
+                    </div>
+                    <div class="stat-number">{{ $conversations->whereIn('status', ['active', 'open'])->count() }}</div>
+                    <div class="stat-label">Đang xử lý</div>
+                    <div class="stat-trend">
+                        {{ $conversations->whereIn('status', ['active', 'open'])->where('updated_at', '>=', now()->subHour())->count() }}
+                        hoạt động gần đây</div>
+                </div>
 
-                                    @if ($isSystem)
-                                        <div class="system-message">
-                                            <i class="fas fa-info-circle me-2"></i>{{ $message->message }}
-                                        </div>
+                <div class="stat-card priority" onclick="filterConversations('priority')">
+                    <div class="stat-header">
+                        <div class="stat-icon">
+                            <i class="fas fa-exclamation-triangle"></i>
+                        </div>
+                    </div>
+                    <div class="stat-number">
+                        {{ $conversations->whereIn('priority', ['high', 'urgent'])->count() }}
+                    </div>
+                    <div class="stat-label">Ưu tiên cao</div>
+                    <div class="stat-trend">Cần xử lý ngay</div>
+                </div>
+
+                <div class="stat-card total" onclick="filterConversations('all')">
+                    <div class="stat-header">
+                        <div class="stat-icon">
+                            <i class="fas fa-chart-line"></i>
+                        </div>
+                    </div>
+                    <div class="stat-number">{{ $conversations->count() }}</div>
+                    <div class="stat-label">Tổng cộng</div>
+                    <div class="stat-trend">{{ $conversations->where('status', 'resolved')->count() }} đã giải quyết</div>
+                </div>
+            </div>
+
+            <!-- Enhanced Chat Interface -->
+            <div class="chat-interface">
+                <!-- Enhanced Sidebar -->
+                <div class="chat-sidebar">
+                    <div class="chat-sidebar-header">
+                        <h3>Danh sách cuộc trò chuyện</h3>
+                        <div class="search-container">
+                            <i class="fas fa-search"></i>
+                            <input type="text" id="search-conversations" placeholder="Tìm kiếm khách hàng...">
+                        </div>
+                    </div>
+                    <div class="conversations-list" id="conversations-list">
+                        @forelse($conversations as $conversation)
+                            @php
+                                $unreadCount = $conversation->messages
+                                    ->where('is_read', false)
+                                    ->where('sender_id', '!=', $user->id)
+                                    ->count();
+                                $customer = $conversation->customer;
+                                $lastMessage = $conversation->messages->last();
+                            @endphp
+                            <div class="conversation-item {{ $loop->first ? 'active' : '' }}"
+                                data-conversation-id="{{ $conversation->id }}" data-status="{{ $conversation->status }}"
+                                data-priority="{{ $conversation->priority ?? 'normal' }}"
+                                data-customer-name="{{ $customer->name ?? 'Khách hàng' }}">
+
+                                <div class="conversation-header">
+                                    <div class="customer-name">
+                                        {{ $customer->name ?? 'Khách hàng' }}
+                                    </div>
+                                    <div class="time">{{ $conversation->updated_at->format('H:i') }}</div>
+                                </div>
+
+                                <div class="message-preview">
+                                    @if ($lastMessage)
+                                        @if ($lastMessage->message)
+                                            {{ Str::limit($lastMessage->message, 45) }}
+                                        @elseif ($lastMessage->attachment)
+                                            📎
+                                            {{ $lastMessage->attachment_type === 'image' ? 'Hình ảnh' : 'Tệp đính kèm' }}
+                                        @else
+                                            Tin nhắn mới
+                                        @endif
                                     @else
-                                        <div class="message {{ $isSent ? 'sent' : 'received' }}">
-                                            <div class="message-info">
-                                                <strong>{{ $isSent ? 'Bạn' : $sender->name ?? 'Khách hàng' }}</strong> •
-                                                {{ $message->created_at->format('H:i') }}
-                                                @if ($isSent)
-                                                    <i class="fas fa-check-double ms-1 text-primary"></i>
-                                                @endif
-                                            </div>
-
-                                            <div class="message-bubble">
-                                                @if ($message->message)
-                                                    <div>{!! nl2br(e($message->message)) !!}</div>
-                                                @endif
-
-                                                @if ($message->attachment)
-                                                    <div class="attachment-preview">
-                                                        @if ($message->attachment_type === 'image')
-                                                            <img src="{{ asset('storage/' . $message->attachment) }}"
-                                                                alt="attachment" class="img-fluid rounded">
-                                                        @else
-                                                            <div class="d-flex align-items-center gap-2">
-                                                                <i class="fas fa-file text-primary"></i>
-                                                                <a href="{{ asset('storage/' . $message->attachment) }}"
-                                                                    class="text-decoration-none" target="_blank">
-                                                                    📎 {{ basename($message->attachment) }}
-                                                                </a>
-                                                            </div>
-                                                        @endif
-                                                    </div>
-                                                @endif
-                                            </div>
-
-
-                                        </div>
+                                        💬 Cuộc trò chuyện mới
                                     @endif
-                                @endforeach
-                            @else
+                                </div>
+
+                                <div class="conversation-meta">
+                                    <span class="status-badge {{ $conversation->status }}">
+                                        @switch($conversation->status)
+                                            @case('distributed')
+                                                📥 Mới nhận
+                                            @break
+
+                                            @case('active')
+                                            @case('open')
+                                                🟢 Đang xử lý
+                                            @break
+
+                                            @case('resolved')
+                                                ✅ Đã giải quyết
+                                            @break
+
+                                            @case('closed')
+                                                🔒 Đã đóng
+                                            @break
+
+                                            @default
+                                                📋 Chờ xử lý
+                                        @endswitch
+                                    </span>
+
+                                    @if ($unreadCount > 0)
+                                        <div class="unread-badge">{{ $unreadCount }}</div>
+                                    @endif
+                                </div>
+                            </div>
+                            @empty
                                 <div class="empty-state">
-                                    <i class="fas fa-comment-dots"></i>
-                                    <h5>Chưa có tin nhắn</h5>
-                                    <p>Hãy bắt đầu cuộc trò chuyện với khách hàng để hỗ trợ họ tốt nhất</p>
+                                    <i class="fas fa-inbox"></i>
+                                    <h5>Không có cuộc trò chuyện</h5>
+                                    <p>Chưa có cuộc trò chuyện nào được phân phối cho chi nhánh này</p>
                                 </div>
-                            @endif
+                            @endforelse
                         </div>
+                    </div>
 
-                        <!-- Enhanced Chat Input -->
-                        <div class="chat-input">
-                            <form id="chat-form">
-                                <div class="input-group">
-                                    <textarea id="message" placeholder="Nhập tin nhắn của bạn..." rows="1"></textarea>
-                                    <div class="actions">
-                                        <button type="button" class="btn-icon" id="attachment-btn" title="Đính kèm tệp">
-                                            <i class="fas fa-paperclip"></i>
-                                        </button>
-                                        <button type="submit" class="btn-icon btn-send" id="send-btn"
-                                            title="Gửi tin nhắn">
-                                            <i class="fas fa-paper-plane"></i>
-                                        </button>
+                    <!-- Enhanced Main Chat Area -->
+                    <div class="chat-main">
+                        @if ($conversations->count() > 0)
+                            @php $firstConversation = $conversations->first(); @endphp
+                            <!-- Enhanced Chat Header -->
+                            <div class="chat-header">
+                                <div class="customer-info">
+                                    <div class="customer-avatar">
+                                        {{ strtoupper(substr($firstConversation->customer->name ?? 'K', 0, 1)) }}
+                                    </div>
+                                    <div class="customer-details">
+                                        <h2>{{ $firstConversation->customer->name ?? 'Khách hàng' }}</h2>
+                                        <div class="email">{{ $firstConversation->customer->email ?? 'Chưa có email' }}</div>
                                     </div>
                                 </div>
-                            </form>
-                            <input type="file" id="attachment" class="d-none"
-                                accept="image/*,.pdf,.doc,.docx,.xls,.xlsx,.txt,.zip,.rar">
-                            <div id="attachment-preview"></div>
-                        </div>
-                    @else
-                        <div class="empty-state">
-                            <i class="fas fa-comments"></i>
-                            <h5>Chưa có cuộc trò chuyện</h5>
-                            <p>Chưa có cuộc trò chuyện nào được phân phối cho chi nhánh này. Hãy chờ admin phân phối hoặc liên
-                                hệ để được hỗ trợ.</p>
-                        </div>
-                    @endif
-                </div>
-            </div>
-        </div>
+                                <div class="chat-actions">
+                                    <div class="dropdown">
+                                        <button class="btn btn-primary dropdown-toggle" type="button" data-bs-toggle="dropdown"
+                                            aria-expanded="false">
+                                            <i class="fas fa-cog me-2"></i>Hành động
+                                        </button>
+                                        <ul class="dropdown-menu dropdown-menu-end">
+                                            <li><a class="dropdown-item" href="#" onclick="updateStatus('active')">
+                                                    <i class="fas fa-play me-2 text-success"></i>Kích hoạt cuộc trò chuyện
+                                                </a></li>
+                                            <li><a class="dropdown-item" href="#" onclick="updateStatus('resolved')">
+                                                    <i class="fas fa-check me-2 text-warning"></i>Đánh dấu đã giải quyết
+                                                </a></li>
+                                            <li><a class="dropdown-item" href="#" onclick="updateStatus('closed')">
+                                                    <i class="fas fa-times me-2 text-danger"></i>Đóng cuộc trò chuyện
+                                                </a></li>
+                                        </ul>
+                                    </div>
+                                </div>
+                            </div>
 
-        {{-- 
-    <!-- COMMENTED OUT: Assign Staff Modal -->
-    <!-- 
-    @if ($user->role === 'branch_manager')
-        <div class="modal fade" id="assignModal" tabindex="-1" aria-hidden="true">
-            <div class="modal-dialog modal-dialog-centered">
-                <div class="modal-content">
-                    <div class="modal-header">
-                        <h5 class="modal-title">
-                            <i class="fas fa-user-plus me-2"></i>Phân công nhân viên
-                        </h5>
-                        <button type="button" class="btn-close" data-bs-dismiss="modal"></button>
-                    </div>
-                    <div class="modal-body">
-                        <form id="assign-form">
-                            <div class="mb-3">
-                                <label for="staff-select" class="form-label">Chọn nhân viên</label>
-                                <select class="form-select" id="staff-select">
-                                    <option value="">Chọn nhân viên...</option>
-                                    @foreach (\App\Models\User::where('branch_id', $user->branch_id)->where('role', 'branch_staff')->get() as $staff)
-                                        <option value="{{ $staff->id }}">{{ $staff->name }} - {{ $staff->email }}</option>
+                            <!-- Enhanced Chat Messages -->
+                            <div class="chat-messages" id="chat-messages">
+                                @if ($firstConversation && $firstConversation->messages->count() > 0)
+                                    @foreach ($firstConversation->messages as $message)
+                                        @php
+                                            $isSent = $message->sender_id === $user->id;
+                                            $sender = $message->sender;
+                                            $isSystem = $message->is_system_message ?? false;
+                                        @endphp
+
+                                        @if ($isSystem)
+                                            <div class="system-message">
+                                                <i class="fas fa-info-circle me-2"></i>{{ $message->message }}
+                                            </div>
+                                        @else
+                                            <div class="message {{ $isSent ? 'sent' : 'received' }}">
+                                                <div class="message-info">
+                                                    <strong>{{ $isSent ? 'Bạn' : $sender->name ?? 'Khách hàng' }}</strong> •
+                                                    {{ $message->created_at->format('H:i') }}
+                                                    @if ($isSent)
+                                                        <i class="fas fa-check-double ms-1 text-primary"></i>
+                                                    @endif
+                                                </div>
+
+                                                <div class="message-bubble">
+                                                    @if ($message->message)
+                                                        <div>{!! nl2br(e($message->message)) !!}</div>
+                                                    @endif
+
+                                                    @if ($message->attachment)
+                                                        <div class="attachment-preview">
+                                                            @if ($message->attachment_type === 'image')
+                                                                <img src="{{ asset('storage/' . $message->attachment) }}"
+                                                                    alt="attachment" class="img-fluid rounded">
+                                                            @else
+                                                                <div class="d-flex align-items-center gap-2">
+                                                                    <i class="fas fa-file text-primary"></i>
+                                                                    <a href="{{ asset('storage/' . $message->attachment) }}"
+                                                                        class="text-decoration-none" target="_blank">
+                                                                        📎 {{ basename($message->attachment) }}
+                                                                    </a>
+                                                                </div>
+                                                            @endif
+                                                        </div>
+                                                    @endif
+                                                </div>
+
+
+                                            </div>
+                                        @endif
                                     @endforeach
-                                </select>
+                                @else
+                                    <div class="empty-state">
+                                        <i class="fas fa-comment-dots"></i>
+                                        <h5>Chưa có tin nhắn</h5>
+                                        <p>Hãy bắt đầu cuộc trò chuyện với khách hàng để hỗ trợ họ tốt nhất</p>
+                                    </div>
+                                @endif
                             </div>
-                            <div class="mb-3">
-                                <label for="assign-note" class="form-label">Ghi chú (tùy chọn)</label>
-                                <textarea class="form-control" id="assign-note" rows="3" 
-                                    placeholder="Thêm ghi chú về việc phân công..."></textarea>
+
+                            <!-- Enhanced Chat Input -->
+                            <div class="chat-input">
+                                <form id="chat-form">
+                                    <div class="input-group">
+                                        <textarea id="message" placeholder="Nhập tin nhắn của bạn..." rows="1"></textarea>
+                                        <div class="actions">
+                                            <button type="button" class="btn-icon" id="image-btn" title="Chọn ảnh">
+                                                <i class="fas fa-image"></i>
+                                            </button>
+                                            <button type="button" class="btn-icon" id="file-btn" title="Chọn tệp">
+                                                <i class="fas fa-paperclip"></i>
+                                            </button>
+                                            <button type="submit" class="btn-icon btn-send" id="send-btn"
+                                                title="Gửi tin nhắn">
+                                                <i class="fas fa-paper-plane"></i>
+                                            </button>
+                                        </div>
+                                    </div>
+                                </form>
+                                <input type="file" id="image" class="d-none" accept="image/*">
+                                <input type="file" id="file" class="d-none"
+                                    accept=".pdf,.doc,.docx,.xls,.xlsx,.txt,.zip,.rar">
+                                <div id="attachment-preview"></div>
                             </div>
-                        </form>
-                    </div>
-                    <div class="modal-footer">
-                        <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">
-                            <i class="fas fa-times me-2"></i>Hủy
-                        </button>
-                        <button type="button" class="btn btn-primary" id="assign-btn">
-                            <i class="fas fa-check me-2"></i>Phân công
-                        </button>
+                        @else
+                            <div class="empty-state">
+                                <i class="fas fa-comments"></i>
+                                <h5>Chưa có cuộc trò chuyện</h5>
+                                <p>Chưa có cuộc trò chuyện nào được phân phối cho chi nhánh này. Hãy chờ admin phân phối hoặc
+                                    liên
+                                    hệ để được hỗ trợ.</p>
+                            </div>
+                        @endif
                     </div>
                 </div>
             </div>
         </div>
-    @endif
-    -->
-    --}}
 
-        <script src="{{ asset('js/chat-realtime.js') }}"></script>
         <script src="https://js.pusher.com/8.2.0/pusher.min.js"></script>
+        <script src="{{ asset('js/echo.js') }}"></script>
+        <script src="{{ asset('js/chat-common.js') }}" defer></script>
         <script>
             document.addEventListener('DOMContentLoaded', function() {
-                console.log('🚀 Branch Chat Interface Loading...');
-
-                // Global variables
-                const currentUserId = {{ $user->id }};
-                let selectedConversationId = {{ $conversations->first()->id ?? 'null' }};
-
-                // DOM elements
-                const chatMessages = document.getElementById('chat-messages');
-                const messageInput = document.getElementById('message');
-                const sendBtn = document.getElementById('send-btn');
-                const attachmentBtn = document.getElementById('attachment-btn');
-                const attachmentInput = document.getElementById('attachment');
-                const attachmentPreview = document.getElementById('attachment-preview');
-                const conversationItems = document.querySelectorAll('.conversation-item');
-                const searchInput = document.getElementById('search-conversations');
-                // const assignBtn = document.getElementById('assign-btn'); // COMMENTED OUT
-
-                console.log('✅ Branch chat initialized with conversation ID:', selectedConversationId);
-
-                // Initialize
-                init();
-
-                function init() {
-                    scrollToBottom();
-                    setupEventListeners();
-                    updateSendButtonState();
-
-                    if (messageInput) {
-                        messageInput.addEventListener('input', autoResizeTextarea);
+                var chatContainer = document.getElementById('chat-container');
+                if (!chatContainer) return;
+                var chat = new ChatCommon({
+                    conversationId: chatContainer.getAttribute('data-conversation-id'),
+                    userId: chatContainer.getAttribute('data-user-id'),
+                    userType: chatContainer.getAttribute('data-user-type'),
+                    api: {
+                        send: '{{ route('branch.chat.send') }}',
+                        getMessages: '/branch/chat/api/conversation/' + chatContainer.getAttribute(
+                            'data-conversation-id'),
+                        typing: '{{ route('branch.chat.typing', [], false) }}'
                     }
-
-                    console.log('✅ Branch chat interface ready');
-                }
-
-                function setupEventListeners() {
-                    console.log('🔧 Setting up event listeners...');
-
-                    // Conversation selection
-                    conversationItems.forEach(item => {
-                        item.addEventListener('click', function() {
-                            const conversationId = this.getAttribute('data-conversation-id');
-                            selectConversation(conversationId);
-                        });
-                    });
-
-                    // Message form submission
-                    const chatForm = document.getElementById('chat-form');
-                    if (chatForm) {
-                        chatForm.addEventListener('submit', function(e) {
-                            e.preventDefault();
-                            sendMessage();
-                        });
-                    }
-
-                    // Enter key to send
-                    if (messageInput) {
-                        messageInput.addEventListener('keypress', function(e) {
-                            if (e.key === 'Enter' && !e.shiftKey) {
-                                e.preventDefault();
-                                sendMessage();
-                            }
-                        });
-
-                        messageInput.addEventListener('input', updateSendButtonState);
-                    }
-
-                    // Attachment button
-                    if (attachmentBtn) {
-                        attachmentBtn.addEventListener('click', function() {
-                            attachmentInput.click();
-                        });
-                    }
-
-                    // File selection
-                    if (attachmentInput) {
-                        attachmentInput.addEventListener('change', handleFileSelection);
-                    }
-
-                    // Search conversations
-                    if (searchInput) {
-                        searchInput.addEventListener('input', function() {
-                            const searchTerm = this.value.toLowerCase();
-                            conversationItems.forEach(item => {
-                                const customerName = item.getAttribute('data-customer-name')
-                                    .toLowerCase();
-                                const messagePreview = item.querySelector('.message-preview')
-                                    .textContent.toLowerCase();
-                                const isVisible = customerName.includes(searchTerm) || messagePreview
-                                    .includes(searchTerm);
-                                item.style.display = isVisible ? '' : 'none';
-                            });
-                        });
-                    }
-
-                    // COMMENTED OUT: Assign staff functionality
-                    /*
-                    if (assignBtn) {
-                        assignBtn.addEventListener('click', assignToStaff);
-                    }
-                    */
-
-                    console.log('✅ Event listeners set up successfully');
-                }
-
-                function scrollToBottom() {
-                    if (chatMessages) {
-                        setTimeout(() => {
-                            chatMessages.scrollTop = chatMessages.scrollHeight;
-                        }, 100);
-                    }
-                }
-
-                function autoResizeTextarea() {
-                    if (messageInput) {
-                        messageInput.style.height = 'auto';
-                        messageInput.style.height = Math.min(messageInput.scrollHeight, 120) + 'px';
-                    }
-                }
-
-                function updateSendButtonState() {
-                    if (messageInput && sendBtn) {
-                        const hasText = messageInput.value.trim().length > 0;
-                        sendBtn.disabled = !hasText;
-                        sendBtn.style.opacity = hasText ? '1' : '0.5';
-                    }
-                }
-
-                function selectConversation(conversationId) {
-                    console.log('🎯 Selecting conversation:', conversationId);
-
-                    selectedConversationId = conversationId;
-
-                    // Update active state with animation
-                    conversationItems.forEach(item => {
-                        item.classList.remove('active');
-                        if (item.getAttribute('data-conversation-id') === conversationId) {
-                            item.classList.add('active');
-                        }
-                    });
-
-                    // Load conversation details
-                    loadConversationDetails(conversationId);
-                }
-
-                async function loadConversationDetails(conversationId) {
-                    try {
-                        // Show loading state
-                        if (chatMessages) {
-                            chatMessages.innerHTML = `
-                            <div class="empty-state">
-                                <div class="loading-spinner"></div>
-                                <h5>Đang tải...</h5>
-                                <p>Vui lòng chờ trong giây lát</p>
-                            </div>
-                        `;
-                        }
-
-                        const response = await fetch(`/branch/chat/api/conversation/${conversationId}`, {
-                            headers: {
-                                'X-CSRF-TOKEN': document.querySelector('meta[name="csrf-token"]')
-                                    .getAttribute('content'),
-                                'X-Requested-With': 'XMLHttpRequest'
-                            }
-                        });
-
-                        const data = await response.json();
-
-                        if (data.success) {
-                            updateChatHeader(data.conversation);
-                            renderMessages(data.conversation.messages || []);
-                        } else {
-                            throw new Error(data.message || 'Failed to load conversation');
-                        }
-                    } catch (error) {
-                        console.error('❌ Error loading conversation:', error);
-                        if (chatMessages) {
-                            chatMessages.innerHTML = `
-                            <div class="empty-state">
-                                <i class="fas fa-exclamation-triangle text-danger"></i>
-                                <h5>Lỗi tải cuộc trò chuyện</h5>
-                                <p>Không thể tải chi tiết cuộc trò chuyện. Vui lòng thử lại.</p>
-                            </div>
-                        `;
-                        }
-                    }
-                }
-
-                function updateChatHeader(conversation) {
-                    const chatHeader = document.querySelector('.chat-header');
-                    if (chatHeader && conversation.customer) {
-                        const customerName = chatHeader.querySelector('.customer-details h2');
-                        const customerEmail = chatHeader.querySelector('.customer-details .email');
-                        const customerAvatar = chatHeader.querySelector('.customer-avatar');
-
-                        if (customerName) customerName.textContent = conversation.customer.name || 'Khách hàng';
-                        if (customerEmail) customerEmail.textContent = conversation.customer.email || 'Chưa có email';
-                        if (customerAvatar) {
-                            customerAvatar.textContent = (conversation.customer.name || 'K').charAt(0).toUpperCase();
-                        }
-                    }
-                }
-
-                function renderMessages(messages) {
-                    if (!chatMessages) return;
-
-                    if (messages.length === 0) {
-                        chatMessages.innerHTML = `
-                        <div class="empty-state">
-                            <i class="fas fa-comment-dots"></i>
-                            <h5>Chưa có tin nhắn</h5>
-                            <p>Hãy bắt đầu cuộc trò chuyện với khách hàng để hỗ trợ họ tốt nhất</p>
-                        </div>
-                    `;
-                        return;
-                    }
-
-                    chatMessages.innerHTML = '';
-                    messages.forEach(message => {
-                        addMessageToChat(message, false);
-                    });
-                    scrollToBottom();
-                }
-
-                async function sendMessage() {
-                    if (!selectedConversationId) {
-                        showNotification('Không có cuộc trò chuyện để gửi tin nhắn', 'error');
-                        return;
-                    }
-
-                    const message = messageInput.value.trim();
-                    const file = attachmentInput.files[0];
-
-                    if (!message && !file) {
-                        showNotification('Vui lòng nhập tin nhắn hoặc chọn tệp', 'warning');
-                        return;
-                    }
-
-                    try {
-                        // Show loading state
-                        sendBtn.disabled = true;
-                        sendBtn.innerHTML = '<div class="loading-spinner"></div>';
-
-                        const formData = new FormData();
-                        formData.append('conversation_id', selectedConversationId);
-                        formData.append('message', message);
-                        if (file) formData.append('attachment', file);
-
-                        const response = await fetch('/branch/chat/send-message', {
-                            method: 'POST',
-                            headers: {
-                                'X-CSRF-TOKEN': document.querySelector('meta[name="csrf-token"]')
-                                    .getAttribute('content'),
-                                'X-Requested-With': 'XMLHttpRequest'
-                            },
-                            body: formData
-                        });
-
-                        const data = await response.json();
-
-                        if (data.success) {
-                            // Clear form
-                            messageInput.value = '';
-                            attachmentInput.value = '';
-                            attachmentPreview.innerHTML = '';
-                            messageInput.style.height = 'auto';
-
-                            // Add message to chat immediately
-                            addMessageToChat({
-                                id: data.data.id,
-                                sender_id: currentUserId,
-                                message: data.data.message,
-                                attachment: data.data.attachment,
-                                attachment_type: data.data.attachment_type,
-                                created_at: data.data.created_at,
-                                sender: {
-                                    name: '{{ $user->name }}'
-                                }
-                            });
-
-                            // Update conversation preview
-                            updateConversationPreview(selectedConversationId, message || '📎 Tệp đính kèm');
-
-                            showNotification('Tin nhắn đã được gửi thành công', 'success');
-                        } else {
-                            throw new Error(data.message || 'Failed to send message');
-                        }
-                    } catch (error) {
-                        console.error('❌ Send message error:', error);
-                        showNotification('Lỗi gửi tin nhắn: ' + error.message, 'error');
-                    } finally {
-                        sendBtn.innerHTML = '<i class="fas fa-paper-plane"></i>';
-                        updateSendButtonState();
-                    }
-                }
-
-                function addMessageToChat(messageData, animate = true) {
-                    if (!chatMessages) return;
-
-                    // Remove empty state if exists
-                    const emptyState = chatMessages.querySelector('.empty-state');
-                    if (emptyState) {
-                        emptyState.remove();
-                    }
-
-                    const isCurrentUser = messageData.sender_id === currentUserId;
-                    const isSystem = messageData.is_system_message ?? false;
-
-                    if (isSystem) {
-                        const systemDiv = document.createElement('div');
-                        systemDiv.className = 'system-message';
-                        if (animate) systemDiv.style.opacity = '0';
-
-                        systemDiv.innerHTML =
-                            `<i class="fas fa-info-circle me-2"></i>${escapeHtml(messageData.message)}`;
-                        chatMessages.appendChild(systemDiv);
-
-                        if (animate) {
-                            setTimeout(() => systemDiv.style.opacity = '1', 100);
-                        }
-                    } else {
-                        const messageDiv = document.createElement('div');
-                        messageDiv.className = `message ${isCurrentUser ? 'sent' : 'received'}`;
-                        if (animate) messageDiv.style.opacity = '0';
-
-                        let messageContent = `<div class="message-bubble">`;
-
-                        if (messageData.message) {
-                            messageContent += `<div>${escapeHtml(messageData.message).replace(/\n/g, '<br>')}</div>`;
-                        }
-
-                        if (messageData.attachment) {
-                            messageContent += `<div class="attachment-preview">`;
-                            if (messageData.attachment_type === 'image') {
-                                messageContent +=
-                                    `<img src="/storage/${messageData.attachment}" alt="attachment" class="img-fluid rounded">`;
-                            } else {
-                                messageContent += `
-                                <div class="d-flex align-items-center gap-2">
-                                    <i class="fas fa-file text-primary"></i>
-                                    <a href="/storage/${messageData.attachment}" class="text-decoration-none" target="_blank">
-                                        📎 ${messageData.attachment.split('/').pop()}
-                                    </a>
-                                </div>
-                            `;
-                            }
-                            messageContent += `</div>`;
-                        }
-
-                        messageContent += `</div>
-                        <div class="message-info">
-                            <strong>${isCurrentUser ? 'Bạn' : (messageData.sender?.name || 'Khách hàng')}</strong> • 
-                            ${new Date(messageData.created_at).toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' })}
-                            ${isCurrentUser ? '<i class="fas fa-check-double ms-1 text-primary"></i>' : ''}
-                        </div>`;
-
-                        messageDiv.innerHTML = messageContent;
-                        chatMessages.appendChild(messageDiv);
-
-                        if (animate) {
-                            setTimeout(() => messageDiv.style.opacity = '1', 100);
-                        }
-                    }
-
-                    scrollToBottom();
-                }
-
-                function updateConversationPreview(conversationId, message) {
-                    const conversationElement = document.querySelector(`[data-conversation-id="${conversationId}"]`);
-                    if (conversationElement) {
-                        const lastMessageElement = conversationElement.querySelector('.message-preview');
-                        if (lastMessageElement) {
-                            lastMessageElement.textContent = message;
-                        }
-                    }
-                }
-
-                function handleFileSelection(e) {
-                    const file = e.target.files[0];
-                    if (!file) {
-                        attachmentPreview.innerHTML = '';
-                        return;
-                    }
-
-                    // Validate file size (10MB max)
-                    if (file.size > 10 * 1024 * 1024) {
-                        showNotification('Tệp quá lớn. Vui lòng chọn tệp nhỏ hơn 10MB', 'error');
-                        e.target.value = '';
-                        return;
-                    }
-
-                    if (file.type.startsWith('image/')) {
-                        const reader = new FileReader();
-                        reader.onload = function(e) {
-                            attachmentPreview.innerHTML = `
-                            <div class="attachment-preview">
-                                <img src="${e.target.result}" alt="Preview">
-                                <div class="file-info">
-                                    <strong>${escapeHtml(file.name)}</strong>
-                                    <div>${(file.size / 1024).toFixed(2)} KB</div>
-                                </div>
-                                <button type="button" class="btn-close" onclick="clearAttachment()">×</button>
-                            </div>
-                        `;
-                        };
-                        reader.readAsDataURL(file);
-                    } else {
-                        attachmentPreview.innerHTML = `
-                        <div class="attachment-preview">
-                            <i class="fas fa-file fa-2x text-primary"></i>
-                            <div class="file-info">
-                                <strong>${escapeHtml(file.name)}</strong>
-                                <div>${(file.size / 1024).toFixed(2)} KB</div>
-                            </div>
-                            <button type="button" class="btn-close" onclick="clearAttachment()">×</button>
-                        </div>
-                    `;
-                    }
-                }
-
-                // Global functions
-                window.clearAttachment = function() {
-                    attachmentInput.value = '';
-                    attachmentPreview.innerHTML = '';
-                };
-
-                window.updateStatus = async function(status) {
-                    if (!selectedConversationId) {
-                        showNotification('Không có cuộc trò chuyện để cập nhật', 'error');
-                        return;
-                    }
-
-                    const statusNames = {
-                        'active': 'Kích hoạt',
-                        'resolved': 'Đã giải quyết',
-                        'closed': 'Đóng'
-                    };
-
-                    if (!confirm(
-                            `Bạn có chắc muốn ${statusNames[status].toLowerCase()} cuộc trò chuyện này?`)) {
-                        return;
-                    }
-
-                    try {
-                        const response = await fetch('/branch/chat/update-status', {
-                            method: 'POST',
-                            headers: {
-                                'Content-Type': 'application/json',
-                                'X-CSRF-TOKEN': document.querySelector('meta[name="csrf-token"]')
-                                    .getAttribute('content'),
-                                'X-Requested-With': 'XMLHttpRequest'
-                            },
-                            body: JSON.stringify({
-                                conversation_id: selectedConversationId,
-                                status: status
-                            })
-                        });
-
-                        const data = await response.json();
-
-                        if (data.success) {
-                            showNotification(`Trạng thái đã được cập nhật thành "${statusNames[status]}"`,
-                                'success');
-                            setTimeout(() => location.reload(), 1500);
-                        } else {
-                            throw new Error(data.message || 'Failed to update status');
-                        }
-                    } catch (error) {
-                        console.error('❌ Update status error:', error);
-                        showNotification('Lỗi cập nhật trạng thái: ' + error.message, 'error');
-                    }
-                };
-
-                // COMMENTED OUT: Staff assignment functionality
-                /*
-                window.assignToStaff = async function() {
-                    const staffSelect = document.getElementById('staff-select');
-                    const assignNote = document.getElementById('assign-note');
-                    
-                    if (!staffSelect || !staffSelect.value) {
-                        showNotification('Vui lòng chọn nhân viên', 'warning');
-                        return;
-                    }
-
-                    try {
-                        const response = await fetch('/branch/chat/assign-staff', {
-                            method: 'POST',
-                            headers: {
-                                'Content-Type': 'application/json',
-                                'X-CSRF-TOKEN': document.querySelector('meta[name="csrf-token"]').getAttribute('content'),
-                                'X-Requested-With': 'XMLHttpRequest'
-                            },
-                            body: JSON.stringify({
-                                conversation_id: selectedConversationId,
-                                staff_id: staffSelect.value,
-                                note: assignNote ? assignNote.value : ''
-                            })
-                        });
-
-                        const data = await response.json();
-
-                        if (data.success) {
-                            showNotification('Phân công nhân viên thành công', 'success');
-                            const modal = bootstrap.Modal.getInstance(document.getElementById('assignModal'));
-                            if (modal) modal.hide();
-                            setTimeout(() => location.reload(), 1500);
-                        } else {
-                            throw new Error(data.message || 'Failed to assign staff');
-                        }
-                    } catch (error) {
-                        console.error('❌ Assign staff error:', error);
-                        showNotification('Lỗi phân công nhân viên: ' + error.message, 'error');
-                    }
-                };
-                */
-
-                window.filterConversations = function(filter) {
-                    console.log('🔍 Filtering conversations by:', filter);
-
-                    conversationItems.forEach(item => {
-                        const status = item.getAttribute('data-status');
-                        const priority = item.getAttribute('data-priority');
-                        let shouldShow = false;
-
-                        switch (filter) {
-                            case 'distributed':
-                                shouldShow = status === 'distributed';
-                                break;
-                            case 'active':
-                                shouldShow = ['active', 'open'].includes(status);
-                                break;
-                            case 'priority':
-                                shouldShow = ['high', 'urgent'].includes(priority);
-                                break;
-                            case 'all':
-                            default:
-                                shouldShow = true;
-                                break;
-                        }
-
-                        item.style.display = shouldShow ? '' : 'none';
-                    });
-
-                    // Update active stat card
-                    document.querySelectorAll('.stat-card').forEach(card => {
-                        card.style.transform = '';
-                        card.style.boxShadow = '';
-                    });
-
-                    const activeCard = document.querySelector(
-                        `.stat-card.${filter === 'active' ? 'active' : filter === 'distributed' ? 'new' : filter === 'priority' ? 'priority' : 'total'}`
-                    );
-                    if (activeCard) {
-                        activeCard.style.transform = 'translateY(-4px)';
-                        activeCard.style.boxShadow = 'var(--shadow-lg)';
-                    }
-                };
-
-                function showNotification(message, type = 'success') {
-                    // Remove existing notifications
-                    document.querySelectorAll('.notification').forEach(n => n.remove());
-
-                    const notification = document.createElement('div');
-                    notification.className = `notification ${type}`;
-                    notification.style.cssText = `
-                    position: fixed;
-                    top: 24px;
-                    right: 24px;
-                    padding: 16px 24px;
-                    border-radius: 12px;
-                    color: white;
-                    font-weight: 600;
-                    z-index: 1000;
-                    transform: translateX(100%);
-                    transition: transform 0.4s ease;
-                    box-shadow: var(--shadow-lg);
-                    min-width: 300px;
-                    max-width: 400px;
-                `;
-
-                    // Set background color based on type
-                    switch (type) {
-                        case 'success':
-                            notification.style.background = 'linear-gradient(135deg, var(--success-color), #059669)';
-                            break;
-                        case 'error':
-                            notification.style.background = 'linear-gradient(135deg, var(--danger-color), #dc2626)';
-                            break;
-                        case 'warning':
-                            notification.style.background = 'linear-gradient(135deg, var (--warning-color), #d97706)';
-                            break;
-                        default:
-                            notification.style.background =
-                                'linear-gradient(135deg, var(--primary-color), var(--primary-hover))';
-                    }
-
-                    notification.textContent = message;
-                    document.body.appendChild(notification);
-
-                    // Show notification
-                    setTimeout(() => notification.style.transform = 'translateX(0)', 100);
-
-                    // Hide notification after 4 seconds
-                    setTimeout(() => {
-                        notification.style.transform = 'translateX(100%)';
-                        setTimeout(() => notification.remove(), 400);
-                    }, 4000);
-                }
-
-                function escapeHtml(text) {
-                    const div = document.createElement('div');
-                    div.textContent = text;
-                    return div.innerHTML;
-                }
-
-                // Khởi tạo Pusher
-                const pusher = new Pusher('{{ config('broadcasting.connections.pusher.key') }}', {
-                    cluster: '{{ config('broadcasting.connections.pusher.options.cluster') }}',
-                    encrypted: true
                 });
-
-                // Lắng nghe tất cả các conversation mà branch có thể thấy
-                @foreach ($conversations as $conv)
-                    pusher.subscribe('chat.{{ $conv->id }}')
-                        .bind('new-message', function(data) {
-                            handleNewMessage(data.message);
-                        });
-                @endforeach
-
-                function handleNewMessage(message) {
-                    const convId = message.conversation_id;
-                    const convItem = document.querySelector(`.conversation-item[data-conversation-id='${convId}']`);
-                    const isCurrent = (window.selectedConversationId == convId); // Cần set biến này khi click vào chat
-
-                    // Nếu đang mở đúng cuộc trò chuyện
-                    if (isCurrent && typeof appendMessageToChat === 'function') {
-                        appendMessageToChat(message);
-                        if (typeof scrollToBottom === 'function') scrollToBottom();
-                    }
-
-                    // Luôn cập nhật preview ở sidebar
-                    if (convItem) {
-                        // Cập nhật preview tin nhắn
-                        const preview = convItem.querySelector('.message-preview');
-                        if (preview) preview.textContent = message.message;
-                        // Cập nhật thời gian
-                        const time = convItem.querySelector('.time');
-                        if (time) time.textContent = formatTime(message.sent_at);
-                        // Tăng badge số chưa đọc nếu chưa mở
-                        if (!isCurrent) {
-                            let badge = convItem.querySelector('.unread-badge');
-                            if (badge) {
-                                badge.textContent = parseInt(badge.textContent || 0) + 1;
-                                badge.style.display = 'inline-block';
-                            } else {
-                                // Nếu chưa có badge, tạo mới
-                                const newBadge = document.createElement('div');
-                                newBadge.className = 'unread-badge';
-                                newBadge.textContent = 1;
-                                convItem.querySelector('.conversation-meta').appendChild(newBadge);
-                            }
-                        }
-                        // Đưa lên đầu danh sách
-                        if (convItem.parentNode.firstChild !== convItem) {
-                            convItem.parentNode.insertBefore(convItem, convItem.parentNode.firstChild);
-                        }
-                    }
-                }
-
-                function formatTime(timeStr) {
-                    const d = new Date(timeStr);
-                    return d.getHours().toString().padStart(2, '0') + ':' + d.getMinutes().toString().padStart(2, '0');
-                }
-
-                console.log('🎉 Branch Chat Interface fully loaded and ready!');
-
             });
         </script>
     @endsection
