+++ resources/views/partials/customer/chat-widget.blade.php
@@ -177,502 +177,26 @@
 <input type="file" id="fileInput" class="hidden" accept=".pdf,.doc,.docx,.txt,.zip,.rar">
 <input type="file" id="imageInput" class="hidden" accept="image/*">
 
+<div id="chat-container" data-conversation-id="{{ $conversation->id ?? '' }}"
+    data-user-id="{{ auth()->id() ?? session('customer_id') }}" data-user-type="customer">
+    <div id="chat-messages" class="chat-messages"></div>
+</div>
+<script src="https://js.pusher.com/8.2.0/pusher.min.js"></script>
+<script src="{{ asset('js/echo.js') }}"></script>
+<script src="{{ asset('js/chat-common.js') }}" defer></script>
 <script>
     document.addEventListener('DOMContentLoaded', function() {
-        // Chat Widget JavaScript
-        const chatToggleBtn = document.getElementById('chatToggleBtn');
-        const chatPopup = document.getElementById('chatPopup');
-        const chatContent = document.getElementById('chatContent');
-        const fullscreenChat = document.getElementById('fullscreenChat');
-        const fullscreenBtn = document.getElementById('fullscreenBtn');
-        const closeChatBtn = document.getElementById('closeChatBtn');
-        const messagesContainer = document.getElementById('messagesContainer');
-        const messageInput = document.getElementById('messageInput');
-        const sendBtn = document.getElementById('sendBtn');
-        const imageBtn = document.getElementById('imageBtn');
-        const fileBtn = document.getElementById('fileBtn');
-        const emojiBtn = document.getElementById('emojiBtn');
-        const emojiPicker = document.getElementById('emojiPicker');
-        const emojiGrid = document.getElementById('emojiGrid');
-        const endChatBtn = document.getElementById('endChatBtn');
-        const ratingModal = document.getElementById('ratingModal');
-        const starRating = document.getElementById('starRating');
-        const feedbackText = document.getElementById('feedbackText');
-        const submitRatingBtn = document.getElementById('submitRatingBtn');
-        const skipRatingBtn = document.getElementById('skipRatingBtn');
-        const successNotification = document.getElementById('successNotification');
-        const closeNotificationBtn = document.getElementById('closeNotificationBtn');
-        const fileInput = document.getElementById('fileInput');
-        const imageInput = document.getElementById('imageInput');
-        const adminStatus = document.getElementById('adminStatus');
-        const adminStatusText = document.getElementById('adminStatusText');
-        const typingIndicator = document.getElementById('typingIndicator');
-        const chatBadge = document.getElementById('chatBadge');
-        const initialTime = document.getElementById('initialTime');
-
-        // State
-        let isChatOpen = false;
-        let isFullscreen = false;
-        let isAdminOnline = true;
-        let isTyping = false;
-        let isChatEnded = false;
-        let currentRating = 0;
-        let messages = [];
-
-        // Emojis
-        const emojis = ['😊', '😂', '❤️', '👍', '👎', '😢', '😮', '😡', '🙏', '👏', '🎉', '🔥'];
-
-        // Initialize
-        init();
-
-        function init() {
-            setupEventListeners();
-            populateEmojis();
-            setupStarRating();
-            setInitialTime();
-
-            // Simulate admin going offline after 5 minutes
-            setTimeout(() => {
-                setAdminOffline();
-            }, 300000);
-        }
-
-        function setupEventListeners() {
-            chatToggleBtn.addEventListener('click', toggleChat);
-            closeChatBtn.addEventListener('click', closeChat);
-            fullscreenBtn.addEventListener('click', toggleFullscreen);
-            sendBtn.addEventListener('click', sendMessage);
-            messageInput.addEventListener('input', handleInputChange);
-            messageInput.addEventListener('keypress', handleKeyPress);
-            imageBtn.addEventListener('click', () => imageInput.click());
-            fileBtn.addEventListener('click', () => fileInput.click());
-            emojiBtn.addEventListener('click', toggleEmojiPicker);
-            endChatBtn.addEventListener('click', endChat);
-            skipRatingBtn.addEventListener('click', closeRatingModal);
-            submitRatingBtn.addEventListener('click', submitRating);
-            closeNotificationBtn.addEventListener('click', closeSuccessNotification);
-            imageInput.addEventListener('change', handleImageUpload);
-            fileInput.addEventListener('change', handleFileUpload);
-
-            // Close emoji picker when clicking outside
-            document.addEventListener('click', (e) => {
-                if (!emojiPicker.contains(e.target) && !emojiBtn.contains(e.target)) {
-                    hideEmojiPicker();
-                }
-            });
-
-            // Auto-resize textarea
-            messageInput.addEventListener('input', autoResizeTextarea);
-        }
-
-        function populateEmojis() {
-            emojiGrid.innerHTML = '';
-            emojis.forEach(emoji => {
-                const button = document.createElement('button');
-                button.textContent = emoji;
-                button.className = 'p-2 hover:bg-gray-100 rounded text-lg transition-colors';
-                button.addEventListener('click', () => addEmoji(emoji));
-                emojiGrid.appendChild(button);
-            });
-        }
-
-        function setupStarRating() {
-            starRating.innerHTML = '';
-            for (let i = 1; i <= 5; i++) {
-                const star = document.createElement('button');
-                star.innerHTML = '<i class="fas fa-star text-2xl"></i>';
-                star.className = 'p-1 transition-colors text-gray-300 hover:text-yellow-300';
-                star.addEventListener('click', () => setRating(i));
-                starRating.appendChild(star);
+        var chatContainer = document.getElementById('chat-container');
+        if (!chatContainer) return;
+        var chat = new ChatCommon({
+            conversationId: chatContainer.getAttribute('data-conversation-id'),
+            userId: chatContainer.getAttribute('data-user-id'),
+            userType: chatContainer.getAttribute('data-user-type'),
+            api: {
+                send: '{{ route('customer.chat.send') }}',
+                getMessages: '/api/conversations/' + chatContainer.getAttribute('data-conversation-id'),
+                typing: '{{ route('customer.chat.typing', [], false) }}'
             }
-        }
-
-        function setInitialTime() {
-            const now = new Date();
-            const timeString = now.toLocaleTimeString('vi-VN', {
-                hour: '2-digit',
-                minute: '2-digit'
-            });
-            initialTime.textContent = timeString;
-        }
-
-        function toggleChat() {
-            console.log('Toggle chat clicked, isChatOpen:', isChatOpen);
-            if (isChatOpen) {
-                closeChat();
-            } else {
-                openChat();
-            }
-        }
-
-        function openChat() {
-            console.log('Opening chat');
-            isChatOpen = true;
-            chatPopup.classList.add('show');
-            console.log('Added show class, chatPopup classes:', chatPopup.className);
-            chatBadge.style.display = 'none';
-            scrollToBottom();
-        }
-
-        function closeChat() {
-            isChatOpen = false;
-            chatPopup.classList.remove('show');
-            if (isFullscreen) {
-                toggleFullscreen();
-            }
-        }
-
-        function toggleFullscreen() {
-            isFullscreen = !isFullscreen;
-            const icon = fullscreenBtn.querySelector('i');
-
-            if (isFullscreen) {
-                // Move to fullscreen
-                fullscreenChat.appendChild(chatContent);
-                fullscreenChat.classList.remove('hidden');
-                chatPopup.style.display = 'none';
-                chatContent.style.height = '100vh';
-                icon.className = 'fas fa-compress text-sm';
-            } else {
-                // Move back to popup
-                chatPopup.appendChild(chatContent);
-                fullscreenChat.classList.add('hidden');
-                chatPopup.style.display = 'block';
-                chatContent.style.height = '600px';
-                icon.className = 'fas fa-expand text-sm';
-            }
-
-            setTimeout(scrollToBottom, 100);
-        }
-
-        function handleInputChange() {
-            const hasText = messageInput.value.trim().length > 0;
-            sendBtn.disabled = !hasText || isChatEnded;
-        }
-
-        function handleKeyPress(e) {
-            if (e.key === 'Enter' && !e.shiftKey) {
-                e.preventDefault();
-                sendMessage();
-            }
-        }
-
-        function autoResizeTextarea() {
-            messageInput.style.height = 'auto';
-            messageInput.style.height = Math.min(messageInput.scrollHeight, 120) + 'px';
-        }
-
-        function sendMessage() {
-            const content = messageInput.value.trim();
-            if (!content || isChatEnded) return;
-
-            const message = {
-                id: Date.now().toString(),
-                content: content,
-                sender: 'user',
-                timestamp: new Date(),
-                type: 'text'
-            };
-
-            addMessage(message);
-            messageInput.value = '';
-            handleInputChange();
-            autoResizeTextarea();
-
-            // Auto-reply
-            if (!isAdminOnline) {
-                setTimeout(() => {
-                    const autoReply = {
-                        id: (Date.now() + 1).toString(),
-                        content: 'Cảm ơn bạn! Admin hiện đang bận, chúng tôi sẽ phản hồi trong vài phút. Bạn vui lòng chờ một chút nhé! 🙏',
-                        sender: 'admin',
-                        timestamp: new Date(),
-                        type: 'text'
-                    };
-                    addMessage(autoReply);
-                }, 1000);
-            } else {
-                // Simulate admin typing
-                showTyping();
-                setTimeout(() => {
-                    hideTyping();
-                    const responses = [
-                        'Cảm ơn bạn đã liên hệ! Tôi sẽ hỗ trợ bạn ngay. 😊',
-                        'Để tôi kiểm tra thông tin cho bạn nhé. 🔍',
-                        'Bạn có thể cho tôi biết thêm chi tiết không? 🤔',
-                        'Tôi hiểu rồi, để tôi hỗ trợ bạn. ✅',
-                        'Bạn có cần tôi gọi điện tư vấn trực tiếp không? 📞'
-                    ];
-                    const randomResponse = responses[Math.floor(Math.random() * responses.length)];
-
-                    const adminReply = {
-                        id: (Date.now() + 1).toString(),
-                        content: randomResponse,
-                        sender: 'admin',
-                        timestamp: new Date(),
-                        type: 'text'
-                    };
-                    addMessage(adminReply);
-                }, 2000);
-            }
-        }
-
-        function addMessage(message) {
-            messages.push(message);
-            const messageElement = createMessageElement(message);
-            messagesContainer.appendChild(messageElement);
-            scrollToBottom();
-        }
-
-        function createMessageElement(message) {
-            const messageDiv = document.createElement('div');
-            messageDiv.className =
-                `flex ${message.sender === 'user' ? 'justify-end' : 'justify-start'} message-enter`;
-
-            const isUser = message.sender === 'user';
-            const timeString = message.timestamp.toLocaleTimeString('vi-VN', {
-                hour: '2-digit',
-                minute: '2-digit'
-            });
-
-            let messageContent = '';
-
-            if (message.type === 'text') {
-                messageContent = `<p class="text-sm whitespace-pre-wrap">${escapeHtml(message.content)}</p>`;
-            } else if (message.type === 'image') {
-                messageContent = `
-                <div class="space-y-2">
-                    <div class="relative group">
-                        <img src="${message.imageUrl}" alt="Uploaded image" 
-                             class="max-w-full h-auto rounded-lg max-h-64 object-cover cursor-pointer hover:opacity-90 transition-opacity"
-                             onclick="window.open('${message.imageUrl}', '_blank')">
-                        <div class="absolute inset-0 bg-black/0 group-hover:bg-black/10 rounded-lg transition-colors flex items-center justify-center">
-                            <div class="opacity-0 group-hover:opacity-100 transition-opacity bg-black/50 text-white px-2 py-1 rounded text-xs">
-                                Click để xem full
-                            </div>
-                        </div>
-                    </div>
-                    ${message.content ? `<p class="text-sm">${escapeHtml(message.content)}</p>` : ''}
-                </div>
-            `;
-            } else if (message.type === 'file') {
-                messageContent = `
-                <div class="flex items-center gap-3 p-3 bg-white/10 rounded-lg min-w-[200px]">
-                    <div class="flex-shrink-0">
-                        <i class="fas fa-paperclip"></i>
-                    </div>
-                    <div class="flex-1 min-w-0">
-                        <p class="text-sm font-medium truncate">${escapeHtml(message.fileName)}</p>
-                        <p class="text-xs opacity-75">${message.fileSize}</p>
-                    </div>
-                    <button class="h-8 w-8 rounded hover:bg-white/20 flex items-center justify-center transition-colors">
-                        <i class="fas fa-download text-sm"></i>
-                    </button>
-                </div>
-            `;
-            }
-
-            messageDiv.innerHTML = `
-            <div class="flex gap-2 max-w-[80%] ${isUser ? 'flex-row-reverse' : 'flex-row'}">
-                <div class="w-8 h-8 ${isUser ? 'bg-blue-500' : 'bg-orange-500'} rounded-full flex items-center justify-center flex-shrink-0">
-                    <span class="text-white text-xs font-bold">${isUser ? 'U' : 'FS'}</span>
-                </div>
-                <div class="flex flex-col ${isUser ? 'items-end' : 'items-start'}">
-                    <div class="rounded-2xl px-4 py-2 max-w-full shadow-sm ${
-                        isUser 
-                            ? 'bg-orange-500 text-white rounded-br-md' 
-                            : 'bg-white text-gray-900 border border-gray-200 rounded-bl-md'
-                    }">
-                        ${messageContent}
-                    </div>
-                    <span class="text-xs text-gray-500 mt-1 px-2">${timeString}</span>
-                </div>
-            </div>
-        `;
-
-            return messageDiv;
-        }
-
-        function handleImageUpload(e) {
-            const file = e.target.files[0];
-            if (!file) return;
-
-            const reader = new FileReader();
-            reader.onload = function(e) {
-                const message = {
-                    id: Date.now().toString(),
-                    content: '',
-                    sender: 'user',
-                    timestamp: new Date(),
-                    type: 'image',
-                    imageUrl: e.target.result
-                };
-                addMessage(message);
-            };
-            reader.readAsDataURL(file);
-
-            // Reset input
-            e.target.value = '';
-        }
-
-        function handleFileUpload(e) {
-            const file = e.target.files[0];
-            if (!file) return;
-
-            const message = {
-                id: Date.now().toString(),
-                content: `Đã gửi file: ${file.name}`,
-                sender: 'user',
-                timestamp: new Date(),
-                type: 'file',
-                fileName: file.name,
-                fileSize: (file.size / 1024 / 1024).toFixed(2) + ' MB',
-                fileUrl: URL.createObjectURL(file)
-            };
-
-            addMessage(message);
-
-            // Reset input
-            e.target.value = '';
-        }
-
-        function toggleEmojiPicker() {
-            if (emojiPicker.classList.contains('show')) {
-                hideEmojiPicker();
-            } else {
-                showEmojiPicker();
-            }
-        }
-
-        function showEmojiPicker() {
-            emojiPicker.classList.remove('hidden');
-            setTimeout(() => emojiPicker.classList.add('show'), 10);
-        }
-
-        function hideEmojiPicker() {
-            emojiPicker.classList.remove('show');
-            setTimeout(() => emojiPicker.classList.add('hidden'), 200);
-        }
-
-        function addEmoji(emoji) {
-            messageInput.value += emoji;
-            handleInputChange();
-            hideEmojiPicker();
-            messageInput.focus();
-        }
-
-        function showTyping() {
-            isTyping = true;
-            typingIndicator.classList.remove('hidden');
-
-            // Add typing animation
-            const typingDiv = document.createElement('div');
-            typingDiv.id = 'typingAnimation';
-            typingDiv.className = 'flex justify-start';
-            typingDiv.innerHTML = `
-            <div class="flex gap-2 max-w-[80%]">
-                <div class="w-8 h-8 bg-orange-500 rounded-full flex items-center justify-center">
-                    <span class="text-white text-xs font-bold">FS</span>
-                </div>
-                <div class="bg-white rounded-2xl rounded-bl-md px-4 py-3 border border-gray-200 shadow-sm">
-                    <div class="flex space-x-1">
-                        <div class="w-2 h-2 bg-gray-400 rounded-full typing-dot"></div>
-                        <div class="w-2 h-2 bg-gray-400 rounded-full typing-dot"></div>
-                        <div class="w-2 h-2 bg-gray-400 rounded-full typing-dot"></div>
-                    </div>
-                </div>
-            </div>
-        `;
-            messagesContainer.appendChild(typingDiv);
-            scrollToBottom();
-        }
-
-        function hideTyping() {
-            isTyping = false;
-            typingIndicator.classList.add('hidden');
-
-            const typingAnimation = document.getElementById('typingAnimation');
-            if (typingAnimation) {
-                typingAnimation.remove();
-            }
-        }
-
-        function setAdminOffline() {
-            isAdminOnline = false;
-            adminStatus.className = 'w-2 h-2 rounded-full bg-gray-400';
-            adminStatusText.textContent = 'Không hoạt động';
-        }
-
-        function endChat() {
-            isChatEnded = true;
-            showRatingModal();
-        }
-
-        function showRatingModal() {
-            ratingModal.classList.remove('hidden');
-        }
-
-        function closeRatingModal() {
-            ratingModal.classList.add('hidden');
-            showSuccessNotification();
-        }
-
-        function setRating(rating) {
-            currentRating = rating;
-            updateStarDisplay();
-            submitRatingBtn.disabled = false;
-        }
-
-        function updateStarDisplay() {
-            const stars = starRating.querySelectorAll('button');
-            stars.forEach((star, index) => {
-                if (index < currentRating) {
-                    star.className = 'p-1 transition-colors text-yellow-400';
-                } else {
-                    star.className = 'p-1 transition-colors text-gray-300 hover:text-yellow-300';
-                }
-            });
-        }
-
-        function submitRating() {
-            const feedback = feedbackText.value.trim();
-
-            // Here you would send the rating to your backend
-            console.log('Rating submitted:', {
-                rating: currentRating,
-                feedback: feedback
-            });
-
-            // You can make an AJAX call here
-            // axios.post('/api/chat/rating', {
-            //     rating: currentRating,
-            //     feedback: feedback
-            // });
-
-            closeRatingModal();
-        }
-
-        function showSuccessNotification() {
-            successNotification.classList.remove('hidden');
-            setTimeout(() => {
-                closeSuccessNotification();
-            }, 5000);
-        }
-
-        function closeSuccessNotification() {
-            successNotification.classList.add('hidden');
-        }
-
-        function scrollToBottom() {
-            setTimeout(() => {
-                messagesContainer.scrollTop = messagesContainer.scrollHeight;
-            }, 100);
-        }
-
-        function escapeHtml(text) {
-            const div = document.createElement('div');
-            div.textContent = text;
-            return div.innerHTML;
-        }
+        });
     });
 </script>
